Move actions tests to separate files
