+++ core/actions/test_test.ts
@@ -1,0 +1,5 @@
+import { suite } from "df/testing";
+
+suite("test", () => {
+  // This action currently has no unit tests!
+});
