+++ core/actions/view_test.ts
@@ -1,0 +1,255 @@
+// tslint:disable tsr-detect-non-literal-fs-filename
+import { expect } from "chai";
+import * as fs from "fs-extra";
+import * as path from "path";
+
+import { asPlainObject, suite, test } from "df/testing";
+import { TmpDirFixture } from "df/testing/fixtures";
+import {
+  coreExecutionRequestFromPath,
+  runMainInVm,
+  VALID_WORKFLOW_SETTINGS_YAML
+} from "df/testing/run_core";
+import {
+  exampleActionDescriptor,
+  exampleBuiltInAssertions,
+  exampleBuiltInAssertionsAsYaml
+} from "df/core/actions/index_test";
+
+suite("view", ({ afterEach }) => {
+  const tmpDirFixture = new TmpDirFixture(afterEach);
+
+  suite("action configs", () => {
+    test(`views can be loaded`, () => {
+      const projectDir = tmpDirFixture.createNewTmpDir();
+      fs.writeFileSync(
+        path.join(projectDir, "workflow_settings.yaml"),
+        VALID_WORKFLOW_SETTINGS_YAML
+      );
+      fs.mkdirSync(path.join(projectDir, "definitions"));
+      fs.writeFileSync(
+        path.join(projectDir, "definitions/actions.yaml"),
+        `
+actions:
+- view:
+    filename: action.sql`
+      );
+      fs.writeFileSync(path.join(projectDir, "definitions/action.sql"), "SELECT 1");
+
+      const result = runMainInVm(coreExecutionRequestFromPath(projectDir));
+
+      expect(result.compile.compiledGraph.graphErrors.compilationErrors).deep.equals([]);
+      expect(asPlainObject(result.compile.compiledGraph.tables)).deep.equals(
+        asPlainObject([
+          {
+            target: {
+              database: "defaultProject",
+              schema: "defaultDataset",
+              name: "action"
+            },
+            canonicalTarget: {
+              database: "defaultProject",
+              schema: "defaultDataset",
+              name: "action"
+            },
+            fileName: "definitions/action.sql",
+            hermeticity: "NON_HERMETIC",
+            query: "SELECT 1",
+            type: "view",
+            enumType: "VIEW",
+            disabled: false
+          }
+        ])
+      );
+    });
+  });
+
+  suite("sqlx and JS API config options", () => {
+    const viewConfig = `{
+    type: "view",
+    name: "name",
+    schema: "dataset",
+    database: "project",
+    dependencies: ["operation"],
+    tags: ["tag1", "tag2"],
+    disabled: true,
+    materialized: true,
+    description: "description",
+    ${exampleActionDescriptor.inputSqlxConfigBlock}
+    bigquery: {
+    labels: {"key": "val"},
+    additionalOptions: {
+        option1Key: "option1",
+        option2Key: "option2",
+    }
+    },
+    dependOnDependencyAssertions: true,
+    hermetic: true,
+    ${exampleBuiltInAssertions.inputAssertionBlock}
+}`;
+    [
+      {
+        filename: "view.sqlx",
+        fileContents: `
+config ${viewConfig}
+SELECT 1`
+      },
+      {
+        filename: "view.js",
+        fileContents: `publish("name", ${viewConfig}).query(ctx => \`\n\nSELECT 1\`)`
+      }
+    ].forEach(testParameters => {
+      test(`for views configured in a ${testParameters.filename} file`, () => {
+        const projectDir = tmpDirFixture.createNewTmpDir();
+        fs.writeFileSync(
+          path.join(projectDir, "workflow_settings.yaml"),
+          VALID_WORKFLOW_SETTINGS_YAML
+        );
+        fs.mkdirSync(path.join(projectDir, "definitions"));
+        fs.writeFileSync(path.join(projectDir, "definitions/operation.sqlx"), "SELECT 1");
+        fs.writeFileSync(
+          path.join(projectDir, `definitions/${testParameters.filename}`),
+          testParameters.fileContents
+        );
+
+        const result = runMainInVm(coreExecutionRequestFromPath(projectDir));
+
+        expect(result.compile.compiledGraph.graphErrors.compilationErrors).deep.equals([]);
+        expect(asPlainObject(result.compile.compiledGraph.tables)).deep.equals([
+          {
+            target: {
+              database: "project",
+              schema: "dataset",
+              name: "name"
+            },
+            canonicalTarget: {
+              database: "project",
+              schema: "dataset",
+              name: "name"
+            },
+            type: "view",
+            disabled: true,
+            hermeticity: "HERMETIC",
+            bigquery: {
+              additionalOptions: {
+                option1Key: "option1",
+                option2Key: "option2"
+              },
+              labels: {
+                key: "val"
+              }
+            },
+            tags: ["tag1", "tag2"],
+            dependencyTargets: [
+              {
+                database: "defaultProject",
+                schema: "defaultDataset",
+                name: "operation"
+              }
+            ],
+            enumType: "VIEW",
+            fileName: `definitions/${testParameters.filename}`,
+            query: "\n\nSELECT 1",
+            actionDescriptor: {
+              ...exampleActionDescriptor.outputActionDescriptor,
+              // sqlxConfig.bigquery.labels are placed as bigqueryLabels.
+              bigqueryLabels: {
+                key: "val"
+              }
+            },
+            materialized: true
+          }
+        ]);
+        expect(asPlainObject(result.compile.compiledGraph.assertions)).deep.equals(
+          exampleBuiltInAssertions.outputAssertions(testParameters.filename)
+        );
+      });
+    });
+  });
+
+  test("action config options", () => {
+    const projectDir = tmpDirFixture.createNewTmpDir();
+    fs.writeFileSync(path.join(projectDir, "workflow_settings.yaml"), VALID_WORKFLOW_SETTINGS_YAML);
+    fs.mkdirSync(path.join(projectDir, "definitions"));
+    fs.writeFileSync(path.join(projectDir, "definitions/operation.sqlx"), "SELECT 1");
+    fs.writeFileSync(path.join(projectDir, "definitions/filename.sql"), "SELECT 1");
+    fs.writeFileSync(
+      path.join(projectDir, "definitions/actions.yaml"),
+      `
+actions:
+- view:
+    name: name
+    dataset: dataset
+    project: project
+    dependencyTargets:
+    - name: operation
+    filename: filename.sql
+    tags:
+    - tag1
+    - tag2
+    disabled: true
+    materialized: true
+    description: description
+    labels:
+      key: val
+    additionalOptions:
+      option1Key: option1
+      option2Key: option2
+    dependOnDependencyAssertions: true
+${exampleBuiltInAssertionsAsYaml.inputActionConfigBlock}
+    hermetic: true
+`
+    );
+
+    const result = runMainInVm(coreExecutionRequestFromPath(projectDir));
+
+    expect(result.compile.compiledGraph.graphErrors.compilationErrors).deep.equals([]);
+    expect(asPlainObject(result.compile.compiledGraph.tables)).deep.equals([
+      {
+        target: {
+          database: "project",
+          schema: "dataset",
+          name: "name"
+        },
+        canonicalTarget: {
+          database: "project",
+          schema: "dataset",
+          name: "name"
+        },
+        type: "view",
+        disabled: true,
+        hermeticity: "HERMETIC",
+        bigquery: {
+          additionalOptions: {
+            option1Key: "option1",
+            option2Key: "option2"
+          },
+          labels: {
+            key: "val"
+          }
+        },
+        tags: ["tag1", "tag2"],
+        dependencyTargets: [
+          {
+            database: "defaultProject",
+            schema: "defaultDataset",
+            name: "operation"
+          }
+        ],
+        enumType: "VIEW",
+        fileName: "definitions/filename.sql",
+        query: "SELECT 1",
+        actionDescriptor: {
+          bigqueryLabels: {
+            key: "val"
+          },
+          description: "description"
+        },
+        materialized: true
+      }
+    ]);
+    expect(asPlainObject(result.compile.compiledGraph.assertions)).deep.equals(
+      exampleBuiltInAssertionsAsYaml.outputAssertions
+    );
+  });
+});
