+++ core/main_test.ts
@@ -3,57 +3,18 @@ import { expect } from "chai";
 import * as fs from "fs-extra";
 import { dump as dumpYaml, load as loadYaml } from "js-yaml";
 import * as path from "path";
-import { CompilerFunction, NodeVM } from "vm2";
 
-import { decode64, encode64 } from "df/common/protos";
-import { compile } from "df/core/compilers";
 import { version } from "df/core/version";
 import { dataform } from "df/protos/ts";
 import { asPlainObject, suite, test } from "df/testing";
+import {
+  coreExecutionRequestFromPath,
+  runMainInVm,
+  VALID_WORKFLOW_SETTINGS_YAML,
+  VALID_DATAFORM_JSON,
+  WorkflowSettingsTemplates
+} from "df/testing/run_core";
 import { TmpDirFixture } from "df/testing/fixtures";
-
-const SOURCE_EXTENSIONS = ["js", "sql", "sqlx", "yaml", "ipynb"];
-
-const VALID_WORKFLOW_SETTINGS_YAML = `
-defaultProject: defaultProject
-defaultDataset: defaultDataset
-defaultLocation: US
-`;
-
-const VALID_DATAFORM_JSON = `
-{
-  "defaultDatabase": "defaultProject",
-  "defaultSchema": "defaultDataset",
-  "defaultLocation": "US"
-}
-`;
-
-class TestConfigs {
-  public static bigquery = dataform.WorkflowSettings.create({
-    defaultDataset: "defaultDataset",
-    defaultLocation: "US"
-  });
-
-  public static bigqueryWithDefaultProject = dataform.WorkflowSettings.create({
-    ...TestConfigs.bigquery,
-    defaultProject: "defaultProject"
-  });
-
-  public static bigqueryWithDatasetSuffix = dataform.WorkflowSettings.create({
-    ...TestConfigs.bigquery,
-    datasetSuffix: "suffix"
-  });
-
-  public static bigqueryWithDefaultProjectAndDataset = dataform.WorkflowSettings.create({
-    ...TestConfigs.bigqueryWithDefaultProject,
-    projectSuffix: "suffix"
-  });
-
-  public static bigqueryWithNamePrefix = dataform.WorkflowSettings.create({
-    ...TestConfigs.bigquery,
-    namePrefix: "prefix"
-  });
-}
 
 const EMPTY_NOTEBOOK_CONTENTS = '{ "cells": [] }';
 
@@ -65,9 +26,9 @@ suite("@dataform/core", ({ afterEach }) => {
   suite("session", () => {
     suite("resolve succeeds", () => {
       [
-        TestConfigs.bigquery,
-        TestConfigs.bigqueryWithDatasetSuffix,
-        TestConfigs.bigqueryWithNamePrefix
+        WorkflowSettingsTemplates.bigquery,
+        WorkflowSettingsTemplates.bigqueryWithDatasetSuffix,
+        WorkflowSettingsTemplates.bigqueryWithNamePrefix
       ].forEach(testConfig => {
         test(`resolve with name prefix "${testConfig.namePrefix}" and dataset suffix "${testConfig.datasetSuffix}"`, () => {
           const projectDir = tmpDirFixture.createNewTmpDir();
@@ -134,9 +95,12 @@ publish("b", {"schema": "foo"}).dependencies("a")`
 
     suite("context methods", () => {
       [
-        TestConfigs.bigqueryWithDefaultProjectAndDataset,
-        { ...TestConfigs.bigqueryWithDatasetSuffix, defaultProject: "defaultProject" },
-        { ...TestConfigs.bigqueryWithNamePrefix, defaultProject: "defaultProject" }
+        WorkflowSettingsTemplates.bigqueryWithDefaultProjectAndDataset,
+        {
+          ...WorkflowSettingsTemplates.bigqueryWithDatasetSuffix,
+          defaultProject: "defaultProject"
+        },
+        { ...WorkflowSettingsTemplates.bigqueryWithNamePrefix, defaultProject: "defaultProject" }
       ].forEach(testConfig => {
         test(
           `assertions target context functions with project suffix '${testConfig.projectSuffix}', ` +
@@ -171,7 +135,7 @@ publish("b", {"schema": "foo"}).dependencies("a")`
         const projectDir = tmpDirFixture.createNewTmpDir();
         fs.writeFileSync(
           path.join(projectDir, "workflow_settings.yaml"),
-          dumpYaml(dataform.WorkflowSettings.create(TestConfigs.bigquery))
+          dumpYaml(dataform.WorkflowSettings.create(WorkflowSettingsTemplates.bigquery))
         );
         fs.mkdirSync(path.join(projectDir, "definitions"));
         fs.writeFileSync(
@@ -327,94 +291,12 @@ publish("b", "SELECT 1;");`
     });
   });
 
-  suite("actions", () => {
-    const getActionsFromResult = (tableType: string, result: dataform.CoreExecutionResponse) => {
-      switch (tableType) {
-        case "table":
-        case "view":
-        case "incremental":
-          return result.compile.compiledGraph.tables;
-        case "operations":
-          return result.compile.compiledGraph.operations;
-        case "assertion":
-          return result.compile.compiledGraph.assertions;
-        default:
-          throw Error(`Unexpected table type: ${tableType}`);
-      }
-    };
-
-    ["table", "view", "incremental", "operations", "assertion"].forEach(tableType => {
-      test(`${tableType} disabled`, () => {
-        const projectDir = tmpDirFixture.createNewTmpDir();
-        fs.writeFileSync(
-          path.join(projectDir, "workflow_settings.yaml"),
-          VALID_WORKFLOW_SETTINGS_YAML
-        );
-        fs.mkdirSync(path.join(projectDir, "definitions"));
-        fs.writeFileSync(
-          path.join(projectDir, `definitions/${tableType}.sqlx`),
-          `config { type: '${tableType}', disabled: true }`
-        );
-
-        const result = runMainInVm(coreExecutionRequestFromPath(projectDir));
-
-        expect(result.compile.compiledGraph.graphErrors.compilationErrors).deep.equals([]);
-        expect(asPlainObject(getActionsFromResult(tableType, result)[0]?.disabled)).equals(true);
-      });
-
-      test(`${tableType} target can be overridden by project config override`, () => {
-        const projectDir = tmpDirFixture.createNewTmpDir();
-        fs.writeFileSync(
-          path.join(projectDir, "workflow_settings.yaml"),
-          VALID_WORKFLOW_SETTINGS_YAML
-        );
-        fs.mkdirSync(path.join(projectDir, "definitions"));
-        fs.writeFileSync(
-          path.join(projectDir, "definitions/file.sqlx"),
-          // If change, then change test "action configs assertions can be loaded".
-          `
-config {
-  type: "${tableType}",
-  name: "name",
-}
-SELECT 1`
-        );
-
-        const result = runMainInVm(
-          coreExecutionRequestFromPath(
-            projectDir,
-            dataform.ProjectConfig.create({
-              defaultDatabase: "otherProject",
-              defaultSchema: "otherDataset",
-              assertionSchema: "otherDataset",
-              tablePrefix: "prefix"
-            })
-          )
-        );
-
-        expect(result.compile.compiledGraph.graphErrors.compilationErrors).deep.equals([]);
-        expect(asPlainObject(getActionsFromResult(tableType, result)[0]?.target)).deep.equals({
-          database: "otherProject",
-          schema: "otherDataset",
-          name: "prefix_name"
-        });
-        expect(
-          asPlainObject(getActionsFromResult(tableType, result)[0]?.canonicalTarget)
-        ).deep.equals({
-          database: "otherProject",
-          schema: "otherDataset",
-          name: "name"
-        });
-      });
-    });
-  });
-
   suite("sqlx special characters", () => {
     test("extract blocks", () => {
       const projectDir = tmpDirFixture.createNewTmpDir();
       fs.writeFileSync(
         path.join(projectDir, "workflow_settings.yaml"),
-        dumpYaml(dataform.WorkflowSettings.create(TestConfigs.bigquery))
+        dumpYaml(dataform.WorkflowSettings.create(WorkflowSettingsTemplates.bigquery))
       );
       fs.mkdirSync(path.join(projectDir, "definitions"));
       fs.writeFileSync(
@@ -464,7 +346,7 @@ SELECT 1`);
       const projectDir = tmpDirFixture.createNewTmpDir();
       fs.writeFileSync(
         path.join(projectDir, "workflow_settings.yaml"),
-        dumpYaml(dataform.WorkflowSettings.create(TestConfigs.bigquery))
+        dumpYaml(dataform.WorkflowSettings.create(WorkflowSettingsTemplates.bigquery))
       );
       fs.mkdirSync(path.join(projectDir, "definitions"));
       const fileContents = `select
@@ -483,7 +365,7 @@ from \`location\``;
       const projectDir = tmpDirFixture.createNewTmpDir();
       fs.writeFileSync(
         path.join(projectDir, "workflow_settings.yaml"),
-        dumpYaml(dataform.WorkflowSettings.create(TestConfigs.bigquery))
+        dumpYaml(dataform.WorkflowSettings.create(WorkflowSettingsTemplates.bigquery))
       );
       fs.mkdirSync(path.join(projectDir, "definitions"));
       const sqlContents = `select
@@ -508,7 +390,7 @@ from \`location\``;
       const projectDir = tmpDirFixture.createNewTmpDir();
       fs.writeFileSync(
         path.join(projectDir, "workflow_settings.yaml"),
-        dumpYaml(dataform.WorkflowSettings.create(TestConfigs.bigquery))
+        dumpYaml(dataform.WorkflowSettings.create(WorkflowSettingsTemplates.bigquery))
       );
       fs.mkdirSync(path.join(projectDir, "definitions"));
       const sqlContents = `select
@@ -948,1237 +830,7 @@ select 1 AS \${dataform.projectConfig.vars.columnVar}`
     });
   });
 
-  suite("notebooks", () => {
-    const createSimpleNotebookProject = (
-      workflowSettingsYaml = VALID_WORKFLOW_SETTINGS_YAML
-    ): string => {
-      const projectDir = tmpDirFixture.createNewTmpDir();
-      fs.writeFileSync(path.join(projectDir, "workflow_settings.yaml"), workflowSettingsYaml);
-      fs.mkdirSync(path.join(projectDir, "definitions"));
-      fs.writeFileSync(
-        path.join(projectDir, "definitions/actions.yaml"),
-        `
-actions:
-- notebook:
-    filename: notebook.ipynb`
-      );
-      return projectDir;
-    };
-
-    test(`notebooks can be loaded via an actions config file`, () => {
-      const projectDir = createSimpleNotebookProject();
-      fs.writeFileSync(
-        path.join(projectDir, "definitions/notebook.ipynb"),
-        EMPTY_NOTEBOOK_CONTENTS
-      );
-
-      const result = runMainInVm(coreExecutionRequestFromPath(projectDir));
-
-      expect(result.compile.compiledGraph.graphErrors.compilationErrors).deep.equals([]);
-      expect(asPlainObject(result.compile.compiledGraph.notebooks)).deep.equals(
-        asPlainObject([
-          {
-            target: {
-              database: "defaultProject",
-              schema: "defaultDataset",
-              name: "notebook"
-            },
-            canonicalTarget: {
-              database: "defaultProject",
-              schema: "defaultDataset",
-              name: "notebook"
-            },
-            fileName: "definitions/notebook.ipynb",
-            notebookContents: JSON.stringify({ cells: [] })
-          }
-        ])
-      );
-    });
-
-    test(`notebook cell output is removed`, () => {
-      const projectDir = createSimpleNotebookProject();
-      fs.writeFileSync(
-        path.join(projectDir, "definitions/notebook.ipynb"),
-        JSON.stringify({
-          cells: [
-            { cell_type: "markdown", source: ["# Some title"], outputs: ["something"] },
-            { cell_type: "code", source: ["print('hi')"], outputs: ["hi"] },
-            { cell_type: "raw", source: ["print('hi')"] }
-          ]
-        })
-      );
-
-      const result = runMainInVm(coreExecutionRequestFromPath(projectDir));
-
-      expect(result.compile.compiledGraph.graphErrors.compilationErrors).deep.equals([]);
-      expect(asPlainObject(result.compile.compiledGraph.notebooks)).deep.equals(
-        asPlainObject([
-          {
-            target: {
-              database: "defaultProject",
-              schema: "defaultDataset",
-              name: "notebook"
-            },
-            canonicalTarget: {
-              database: "defaultProject",
-              schema: "defaultDataset",
-              name: "notebook"
-            },
-            fileName: "definitions/notebook.ipynb",
-            notebookContents: JSON.stringify({
-              cells: [
-                { cell_type: "markdown", source: ["# Some title"], outputs: [] },
-                { cell_type: "code", source: ["print('hi')"], outputs: [] },
-                { cell_type: "raw", source: ["print('hi')"] }
-              ]
-            })
-          }
-        ])
-      );
-    });
-
-    test(`notebook default runtime options are loaded`, () => {
-      const projectDir = createSimpleNotebookProject(`
-defaultProject: dataform
-defaultLocation: US
-defaultNotebookRuntimeOptions:
-  outputBucket: gs://some-bucket
-  runtimeTemplateName: projects/test-project/locations/us-central1/notebookRuntimeTemplates/test-template
-  `);
-      fs.writeFileSync(
-        path.join(projectDir, "definitions/notebook.ipynb"),
-        EMPTY_NOTEBOOK_CONTENTS
-      );
-
-      const result = runMainInVm(coreExecutionRequestFromPath(projectDir));
-
-      expect(result.compile.compiledGraph.graphErrors.compilationErrors).deep.equals([]);
-      expect(asPlainObject(result.compile.compiledGraph.projectConfig)).deep.equals({
-        defaultDatabase: "dataform",
-        defaultLocation: "US",
-        defaultNotebookRuntimeOptions: {
-          outputBucket: "gs://some-bucket",
-          runtimeTemplateName:
-            "projects/test-project/locations/us-central1/notebookRuntimeTemplates/test-template"
-        },
-        warehouse: "bigquery"
-      });
-    });
-  });
-
-  suite("data preparations", () => {
-    const createSimpleDataPreparationProject = (
-      workflowSettingsYaml = VALID_WORKFLOW_SETTINGS_YAML,
-      writeActionsYaml = true
-    ): string => {
-      const projectDir = tmpDirFixture.createNewTmpDir();
-      fs.writeFileSync(path.join(projectDir, "workflow_settings.yaml"), workflowSettingsYaml);
-      fs.mkdirSync(path.join(projectDir, "definitions"));
-
-      if (writeActionsYaml) {
-        fs.writeFileSync(
-          path.join(projectDir, "definitions/actions.yaml"),
-          `
-  actions:
-  - dataPreparation:
-      filename: data_preparation.dp.yaml`
-        );
-      }
-      return projectDir;
-    };
-
-    test(`empty data preparation returns a default target`, () => {
-      const projectDir = createSimpleDataPreparationProject();
-      const dataPreparationYaml = `
-`;
-
-      fs.writeFileSync(
-        path.join(projectDir, "definitions/data_preparation.dp.yaml"),
-        dataPreparationYaml
-      );
-
-      const result = runMainInVm(coreExecutionRequestFromPath(projectDir));
-
-      expect(result.compile.compiledGraph.graphErrors.compilationErrors).deep.equals([]);
-      expect(asPlainObject(result.compile.compiledGraph.dataPreparations)).deep.equals(
-        asPlainObject([
-          {
-            target: {
-              database: "defaultProject",
-              schema: "defaultDataset",
-              name: "data_preparation"
-            },
-            canonicalTarget: {
-              database: "defaultProject",
-              schema: "defaultDataset",
-              name: "data_preparation"
-            },
-            targets: [
-              {
-                database: "defaultProject",
-                schema: "defaultDataset",
-                name: "data_preparation"
-              }
-            ],
-            canonicalTargets: [
-              {
-                database: "defaultProject",
-                schema: "defaultDataset",
-                name: "data_preparation"
-              }
-            ],
-            fileName: "definitions/data_preparation.dp.yaml",
-            dataPreparationYaml: ""
-          }
-        ])
-      );
-    });
-
-    test(`data preparation with no targets a default target`, () => {
-      const projectDir = createSimpleDataPreparationProject();
-      const dataPreparationYaml = `
-nodes:
-- id: node1
-  source:
-    table:
-      project: prj
-      dataset: ds
-      table: src
-  generated:
-    outputSchema:
-      field:
-      - name: a
-        type: INT64
-        mode: NULLABLE
-    sourceGenerated:
-      sourceSchema:
-        tableSchema:
-          field:
-          - name: a
-            type: STRING
-            mode: NULLABLE
-`;
-
-      fs.writeFileSync(
-        path.join(projectDir, "definitions/data_preparation.dp.yaml"),
-        dataPreparationYaml
-      );
-
-      const result = runMainInVm(coreExecutionRequestFromPath(projectDir));
-
-      expect(result.compile.compiledGraph.graphErrors.compilationErrors).deep.equals([]);
-      expect(asPlainObject(result.compile.compiledGraph.dataPreparations)).deep.equals(
-        asPlainObject([
-          {
-            target: {
-              database: "defaultProject",
-              schema: "defaultDataset",
-              name: "data_preparation"
-            },
-            canonicalTarget: {
-              database: "defaultProject",
-              schema: "defaultDataset",
-              name: "data_preparation"
-            },
-            targets: [
-              {
-                database: "defaultProject",
-                schema: "defaultDataset",
-                name: "data_preparation"
-              }
-            ],
-            canonicalTargets: [
-              {
-                database: "defaultProject",
-                schema: "defaultDataset",
-                name: "data_preparation"
-              }
-            ],
-            fileName: "definitions/data_preparation.dp.yaml",
-            dataPreparationYaml: dumpYaml(loadYaml(dataPreparationYaml))
-          }
-        ])
-      );
-    });
-
-    test(`data preparations can be loaded via sqlx file`, () => {
-      const projectDir = createSimpleDataPreparationProject(VALID_WORKFLOW_SETTINGS_YAML, false);
-      const dataPreparationSqlx = `
-config {
-  type: "dataPreparation",
-  name: "dest",
-  dataset: "ds",
-  project: "prj",
-  errorTable: {
-    name: "errorTable",
-    dataset: "errorDs",
-    project: "errorPrj",
-  }
-}
-
-FROM x
--- Ensure y is positive
-|> $\{EXPECT\} y > 0
-$\{when(true, "|> SELECT *", "|> SELECT 1")\}
-`;
-
-      fs.writeFileSync(
-        path.join(projectDir, "definitions/data_preparation.sqlx"),
-        dataPreparationSqlx
-      );
-
-      const result = runMainInVm(coreExecutionRequestFromPath(projectDir));
-
-      expect(result.compile.compiledGraph.graphErrors.compilationErrors).deep.equals([]);
-      expect(asPlainObject(result.compile.compiledGraph.dataPreparations)).deep.equals(
-        asPlainObject([
-          {
-            target: {
-              database: "prj",
-              schema: "ds",
-              name: "dest"
-            },
-            canonicalTarget: {
-              database: "prj",
-              schema: "ds",
-              name: "dest"
-            },
-            targets: [
-              {
-                database: "prj",
-                schema: "ds",
-                name: "dest"
-              },
-              {
-                database: "errorPrj",
-                schema: "errorDs",
-                name: "errorTable"
-              }
-            ],
-            canonicalTargets: [
-              {
-                database: "prj",
-                schema: "ds",
-                name: "dest"
-              },
-              {
-                database: "errorPrj",
-                schema: "errorDs",
-                name: "errorTable"
-              }
-            ],
-            fileName: "definitions/data_preparation.sqlx",
-            load: {
-              replace: {}
-            },
-            query: `FROM x
--- Ensure y is positive
-|>  /* @@VALIDATION */ WHERE  y > 0
-|> SELECT *`,
-            errorTable: {
-              database: "errorPrj",
-              schema: "errorDs",
-              name: "errorTable"
-            },
-            errorTableRetentionDays: 0
-          }
-        ])
-      );
-    });
-
-    test(`data preparations can be loaded via sqlx file with file name as action name`, () => {
-      const projectDir = createSimpleDataPreparationProject(VALID_WORKFLOW_SETTINGS_YAML, false);
-      // TODO(fernst): decide on future of `validate(`.
-      const dataPreparationSqlx = `
-config {
-  type: "dataPreparation",
-  dataset: "ds",
-  project: "prj",
-  errorTable: {
-    name: "errorTable",
-    dataset: "errorDs",
-    project: "errorPrj",
-  }
-}
-
-FROM x
-$\{when(true, "|> SELECT *", "|> SELECT 1")\}
-`;
-
-      fs.writeFileSync(
-        path.join(projectDir, "definitions/this_is_the_file_name.sqlx"),
-        dataPreparationSqlx
-      );
-
-      const result = runMainInVm(coreExecutionRequestFromPath(projectDir));
-
-      expect(result.compile.compiledGraph.graphErrors.compilationErrors).deep.equals([]);
-      expect(asPlainObject(result.compile.compiledGraph.dataPreparations)).deep.equals(
-        asPlainObject([
-          {
-            target: {
-              database: "prj",
-              schema: "ds",
-              name: "this_is_the_file_name"
-            },
-            canonicalTarget: {
-              database: "prj",
-              schema: "ds",
-              name: "this_is_the_file_name"
-            },
-            targets: [
-              {
-                database: "prj",
-                schema: "ds",
-                name: "this_is_the_file_name"
-              },
-              {
-                database: "errorPrj",
-                schema: "errorDs",
-                name: "errorTable"
-              }
-            ],
-            canonicalTargets: [
-              {
-                database: "prj",
-                schema: "ds",
-                name: "this_is_the_file_name"
-              },
-              {
-                database: "errorPrj",
-                schema: "errorDs",
-                name: "errorTable"
-              }
-            ],
-            fileName: "definitions/this_is_the_file_name.sqlx",
-            load: {
-              replace: {}
-            },
-            query: `FROM x
-|> SELECT *`,
-            errorTable: {
-              database: "errorPrj",
-              schema: "errorDs",
-              name: "errorTable"
-            },
-            errorTableRetentionDays: 0
-          }
-        ])
-      );
-    });
-
-    test(`data preparations can be loaded via dp.sqlx file with file name as action name`, () => {
-      const projectDir = createSimpleDataPreparationProject(VALID_WORKFLOW_SETTINGS_YAML, false);
-      // TODO(fernst): decide on future of `validate(`.
-      const dataPreparationSqlx = `
-config {
-  type: "dataPreparation",
-  dataset: "ds",
-  project: "prj",
-  errorTable: {
-    name: "errorTable",
-    dataset: "errorDs",
-    project: "errorPrj",
-  }
-}
-
-FROM x
-$\{when(true, "|> SELECT *", "|> SELECT 1")\}
-`;
-
-      fs.writeFileSync(
-        path.join(projectDir, "definitions/this_is_the_file_name.dp.sqlx"),
-        dataPreparationSqlx
-      );
-
-      const result = runMainInVm(coreExecutionRequestFromPath(projectDir));
-
-      expect(result.compile.compiledGraph.graphErrors.compilationErrors).deep.equals([]);
-      expect(asPlainObject(result.compile.compiledGraph.dataPreparations)).deep.equals(
-        asPlainObject([
-          {
-            target: {
-              database: "prj",
-              schema: "ds",
-              name: "this_is_the_file_name"
-            },
-            canonicalTarget: {
-              database: "prj",
-              schema: "ds",
-              name: "this_is_the_file_name"
-            },
-            targets: [
-              {
-                database: "prj",
-                schema: "ds",
-                name: "this_is_the_file_name"
-              },
-              {
-                database: "errorPrj",
-                schema: "errorDs",
-                name: "errorTable"
-              }
-            ],
-            canonicalTargets: [
-              {
-                database: "prj",
-                schema: "ds",
-                name: "this_is_the_file_name"
-              },
-              {
-                database: "errorPrj",
-                schema: "errorDs",
-                name: "errorTable"
-              }
-            ],
-            fileName: "definitions/this_is_the_file_name.dp.sqlx",
-            load: {
-              replace: {}
-            },
-            query: `FROM x
-|> SELECT *`,
-            errorTable: {
-              database: "errorPrj",
-              schema: "errorDs",
-              name: "errorTable"
-            },
-            errorTableRetentionDays: 0
-          }
-        ])
-      );
-    });
-
-    test(`data preparations can be loaded via sqlx file with compilation overrides`, () => {
-      const projectDir = createSimpleDataPreparationProject(VALID_WORKFLOW_SETTINGS_YAML, false);
-      const dataPreparationSqlx = `
-config {
-  type: "dataPreparation",
-  name: "dest",
-  errorTable: {
-    name: "errorTable",
-  },
-  loadMode: {
-    mode: "APPEND",
-  },
-}
-
-FROM x
-|> SELECT *
-`;
-
-      fs.writeFileSync(
-        path.join(projectDir, "definitions/data_preparation.sqlx"),
-        dataPreparationSqlx
-      );
-
-      const coreExecutionRequest = dataform.CoreExecutionRequest.create({
-        compile: {
-          compileConfig: {
-            projectDir,
-            filePaths: ["definitions/data_preparation.sqlx"],
-            projectConfigOverride: {
-              defaultDatabase: "projectOverride",
-              defaultSchema: "datasetOverride"
-            }
-          }
-        }
-      });
-
-      const result = runMainInVm(coreExecutionRequest);
-
-      expect(result.compile.compiledGraph.graphErrors.compilationErrors).deep.equals([]);
-      expect(asPlainObject(result.compile.compiledGraph.dataPreparations)).deep.equals(
-        asPlainObject([
-          {
-            target: {
-              database: "projectOverride",
-              schema: "datasetOverride",
-              name: "dest"
-            },
-            canonicalTarget: {
-              database: "projectOverride",
-              schema: "datasetOverride",
-              name: "dest"
-            },
-            targets: [
-              {
-                database: "projectOverride",
-                schema: "datasetOverride",
-                name: "dest"
-              },
-              {
-                database: "projectOverride",
-                schema: "datasetOverride",
-                name: "errorTable"
-              }
-            ],
-            canonicalTargets: [
-              {
-                database: "projectOverride",
-                schema: "datasetOverride",
-                name: "dest"
-              },
-              {
-                database: "projectOverride",
-                schema: "datasetOverride",
-                name: "errorTable"
-              }
-            ],
-            fileName: "definitions/data_preparation.sqlx",
-            query: "FROM x\n|> SELECT *",
-            load: {
-              append: {}
-            },
-            errorTable: {
-              database: "projectOverride",
-              schema: "datasetOverride",
-              name: "errorTable"
-            },
-            errorTableRetentionDays: 0
-          }
-        ])
-      );
-    });
-
-    test(`data preparations can be loaded via sqlx file with project defaults`, () => {
-      const projectDir = createSimpleDataPreparationProject(VALID_WORKFLOW_SETTINGS_YAML, false);
-      const dataPreparationSqlx = `
-config {
-  type: "dataPreparation",
-  name: "dest",
-  errorTable: {
-    name: "errorTable",
-  },
-  loadMode: {
-    mode: "MAXIMUM",
-    incrementalColumn: "xyz",
-  },
-}
-
-FROM x
-|> SELECT *
-`;
-
-      fs.writeFileSync(
-        path.join(projectDir, "definitions/data_preparation.sqlx"),
-        dataPreparationSqlx
-      );
-
-      const result = runMainInVm(coreExecutionRequestFromPath(projectDir));
-
-      expect(result.compile.compiledGraph.graphErrors.compilationErrors).deep.equals([]);
-      expect(asPlainObject(result.compile.compiledGraph.dataPreparations)).deep.equals(
-        asPlainObject([
-          {
-            target: {
-              database: "defaultProject",
-              schema: "defaultDataset",
-              name: "dest"
-            },
-            canonicalTarget: {
-              database: "defaultProject",
-              schema: "defaultDataset",
-              name: "dest"
-            },
-            targets: [
-              {
-                database: "defaultProject",
-                schema: "defaultDataset",
-                name: "dest"
-              },
-              {
-                database: "defaultProject",
-                schema: "defaultDataset",
-                name: "errorTable"
-              }
-            ],
-            canonicalTargets: [
-              {
-                database: "defaultProject",
-                schema: "defaultDataset",
-                name: "dest"
-              },
-              {
-                database: "defaultProject",
-                schema: "defaultDataset",
-                name: "errorTable"
-              }
-            ],
-            fileName: "definitions/data_preparation.sqlx",
-            load: {
-              maximum: {
-                columnName: "xyz"
-              }
-            },
-            query: "FROM x\n|> SELECT *",
-            errorTable: {
-              database: "defaultProject",
-              schema: "defaultDataset",
-              name: "errorTable"
-            },
-            errorTableRetentionDays: 0
-          }
-        ])
-      );
-    });
-
-    test(`data preparations can be loaded via an actions config file`, () => {
-      const projectDir = createSimpleDataPreparationProject();
-      const dataPreparationYaml = `
-nodes:
-- id: node1
-  source:
-    table:
-      project: prj
-      dataset: ds
-      table: src
-  destination:
-    table:
-      project: prj
-      dataset: ds
-      table: dest
-  generated:
-    outputSchema:
-      field:
-      - name: a
-        type: INT64
-        mode: NULLABLE
-    sourceGenerated:
-      sourceSchema:
-        tableSchema:
-          field:
-          - name: a
-            type: STRING
-            mode: NULLABLE
-    destinationGenerated:
-      schema:
-        field:
-        - name: a
-          type: STRING
-          mode: NULLABLE
-`;
-
-      fs.writeFileSync(
-        path.join(projectDir, "definitions/data_preparation.dp.yaml"),
-        dataPreparationYaml
-      );
-
-      const result = runMainInVm(coreExecutionRequestFromPath(projectDir));
-
-      expect(result.compile.compiledGraph.graphErrors.compilationErrors).deep.equals([]);
-      expect(asPlainObject(result.compile.compiledGraph.dataPreparations)).deep.equals(
-        asPlainObject([
-          {
-            target: {
-              database: "prj",
-              schema: "ds",
-              name: "dest"
-            },
-            canonicalTarget: {
-              database: "prj",
-              schema: "ds",
-              name: "dest"
-            },
-            targets: [
-              {
-                database: "prj",
-                schema: "ds",
-                name: "dest"
-              }
-            ],
-            canonicalTargets: [
-              {
-                database: "prj",
-                schema: "ds",
-                name: "dest"
-              }
-            ],
-            fileName: "definitions/data_preparation.dp.yaml",
-            dataPreparationYaml: dumpYaml(loadYaml(dataPreparationYaml))
-          }
-        ])
-      );
-    });
-
-    test(`data preparations resolves compilation overrides before encoding`, () => {
-      const projectDir = createSimpleDataPreparationProject(`
-defaultProject: defaultProject
-defaultDataset: defaultDataset
-defaultLocation: US
-projectSuffix: projectSuffix
-datasetSuffix: datasetSuffix
-namePrefix: tablePrefix
-`);
-      const dataPreparationYaml = `
-configuration:
-  errorTable:
-    table: error
-nodes:
-- id: node1
-  source:
-    table:
-      table: src
-  generated:
-    sourceGenerated:
-      sourceSchema:
-        tableSchema:
-          field:
-          - name: a
-            type: STRING
-            mode: NULLABLE
-    outputSchema:
-      field:
-      - name: a
-        type: INT64
-        mode: NULLABLE
-- id: node2
-  source:
-    nodeId: node1
-  destination:
-    table:
-      table: dest
-  generated:
-    sourceGenerated:
-      sourceSchema:
-        nodeSchema:
-          field:
-          - name: a
-            type: STRING
-            mode: NULLABLE
-    outputSchema:
-      field:
-      - name: a
-        type: INT64
-        mode: NULLABLE
-    destinationGenerated:
-      schema:
-        field:
-        - name: a
-          type: STRING
-          mode: NULLABLE
-`;
-
-      fs.writeFileSync(
-        path.join(projectDir, "definitions/data_preparation.dp.yaml"),
-        dataPreparationYaml
-      );
-
-      const resolvedYaml = `
-configuration:
-  errorTable:
-    project: defaultProject_projectSuffix
-    dataset: defaultDataset_datasetSuffix
-    table: tablePrefix_error
-nodes:
-- id: node1
-  source:
-    table:
-      project: defaultProject_projectSuffix
-      dataset: defaultDataset_datasetSuffix
-      table: tablePrefix_src
-  generated:
-    sourceGenerated:
-      sourceSchema:
-        tableSchema:
-          field:
-          - name: a
-            type: STRING
-            mode: NULLABLE
-    outputSchema:
-      field:
-      - name: a
-        type: INT64
-        mode: NULLABLE
-- id: node2
-  source:
-    nodeId: node1
-  destination:
-    table:
-      project: defaultProject_projectSuffix
-      dataset: defaultDataset_datasetSuffix
-      table: tablePrefix_dest
-  generated:
-    sourceGenerated:
-      sourceSchema:
-        nodeSchema:
-          field:
-          - name: a
-            type: STRING
-            mode: NULLABLE
-    outputSchema:
-      field:
-      - name: a
-        type: INT64
-        mode: NULLABLE
-    destinationGenerated:
-      schema:
-        field:
-        - name: a
-          type: STRING
-          mode: NULLABLE
-`;
-
-      const result = runMainInVm(coreExecutionRequestFromPath(projectDir));
-
-      expect(result.compile.compiledGraph.graphErrors.compilationErrors).deep.equals([]);
-      expect(asPlainObject(result.compile.compiledGraph.dataPreparations)).deep.equals(
-        asPlainObject([
-          {
-            target: {
-              database: "defaultProject_projectSuffix",
-              schema: "defaultDataset_datasetSuffix",
-              name: "tablePrefix_dest"
-            },
-            canonicalTarget: {
-              database: "defaultProject",
-              schema: "defaultDataset",
-              name: "dest"
-            },
-            targets: [
-              {
-                database: "defaultProject_projectSuffix",
-                schema: "defaultDataset_datasetSuffix",
-                name: "tablePrefix_dest"
-              }
-            ],
-            canonicalTargets: [
-              {
-                database: "defaultProject",
-                schema: "defaultDataset",
-                name: "dest"
-              }
-            ],
-            fileName: "definitions/data_preparation.dp.yaml",
-            dataPreparationYaml: dumpYaml(loadYaml(resolvedYaml))
-          }
-        ])
-      );
-    });
-  });
-
   suite("action configs", () => {
-    test(`operations can be loaded`, () => {
-      const projectDir = tmpDirFixture.createNewTmpDir();
-      fs.writeFileSync(
-        path.join(projectDir, "workflow_settings.yaml"),
-        VALID_WORKFLOW_SETTINGS_YAML
-      );
-      fs.mkdirSync(path.join(projectDir, "definitions"));
-      fs.writeFileSync(
-        path.join(projectDir, "definitions/actions.yaml"),
-        `
-actions:
-- operation:
-    filename: action.sql`
-      );
-      fs.writeFileSync(path.join(projectDir, "definitions/action.sql"), "SELECT 1");
-
-      const result = runMainInVm(coreExecutionRequestFromPath(projectDir));
-
-      expect(result.compile.compiledGraph.graphErrors.compilationErrors).deep.equals([]);
-      expect(asPlainObject(result.compile.compiledGraph.operations)).deep.equals(
-        asPlainObject([
-          {
-            target: {
-              database: "defaultProject",
-              schema: "defaultDataset",
-              name: "action"
-            },
-            canonicalTarget: {
-              database: "defaultProject",
-              schema: "defaultDataset",
-              name: "action"
-            },
-            fileName: "definitions/action.sql",
-            queries: ["SELECT 1"],
-            hermeticity: "NON_HERMETIC"
-          }
-        ])
-      );
-    });
-
-    test(`declarations can be loaded`, () => {
-      const projectDir = tmpDirFixture.createNewTmpDir();
-      fs.writeFileSync(
-        path.join(projectDir, "workflow_settings.yaml"),
-        VALID_WORKFLOW_SETTINGS_YAML
-      );
-      fs.mkdirSync(path.join(projectDir, "definitions"));
-      fs.writeFileSync(
-        path.join(projectDir, "definitions/actions.yaml"),
-        `
-actions:
-- declaration:
-    name: action`
-      );
-
-      const result = runMainInVm(coreExecutionRequestFromPath(projectDir));
-
-      expect(result.compile.compiledGraph.graphErrors.compilationErrors).deep.equals([]);
-      expect(asPlainObject(result.compile.compiledGraph.declarations)).deep.equals(
-        asPlainObject([
-          {
-            target: {
-              database: "defaultProject",
-              schema: "defaultDataset",
-              name: "action"
-            },
-            canonicalTarget: {
-              database: "defaultProject",
-              schema: "defaultDataset",
-              name: "action"
-            }
-          }
-        ])
-      );
-    });
-
-    test(`fails when filename is defined for declaration`, () => {
-      const projectDir = tmpDirFixture.createNewTmpDir();
-      fs.writeFileSync(
-        path.join(projectDir, "workflow_settings.yaml"),
-        VALID_WORKFLOW_SETTINGS_YAML
-      );
-      fs.mkdirSync(path.join(projectDir, "definitions"));
-      fs.writeFileSync(
-        path.join(projectDir, "definitions/actions.yaml"),
-        `
-actions:
-- declaration:
-    fileName: doesnotexist.sql
-    name: name`
-      );
-
-      expect(() => runMainInVm(coreExecutionRequestFromPath(projectDir))).to.throw(
-        `Unexpected property "fileName", or property value type of "string" is incorrect. See https://dataform-co.github.io/dataform/docs/configs-reference#dataform-ActionConfigs for allowed properties.`
-      );
-    });
-
-    test(`fails when target name is not defined for declaration`, () => {
-      const projectDir = tmpDirFixture.createNewTmpDir();
-      fs.writeFileSync(
-        path.join(projectDir, "workflow_settings.yaml"),
-        VALID_WORKFLOW_SETTINGS_YAML
-      );
-      fs.mkdirSync(path.join(projectDir, "definitions"));
-      fs.writeFileSync(
-        path.join(projectDir, "definitions/actions.yaml"),
-        `
-actions:
-- declaration:
-    dataset: test`
-      );
-
-      expect(() => runMainInVm(coreExecutionRequestFromPath(projectDir))).to.throw(
-        "Declarations must have a populated 'name' field."
-      );
-    });
-
-    test(`tables can be loaded`, () => {
-      const projectDir = tmpDirFixture.createNewTmpDir();
-      fs.writeFileSync(
-        path.join(projectDir, "workflow_settings.yaml"),
-        VALID_WORKFLOW_SETTINGS_YAML
-      );
-      fs.mkdirSync(path.join(projectDir, "definitions"));
-      fs.writeFileSync(
-        path.join(projectDir, "definitions/actions.yaml"),
-        `
-actions:
-- table:
-    filename: action.sql`
-      );
-      fs.writeFileSync(path.join(projectDir, "definitions/action.sql"), "SELECT 1");
-
-      const result = runMainInVm(coreExecutionRequestFromPath(projectDir));
-
-      expect(result.compile.compiledGraph.graphErrors.compilationErrors).deep.equals([]);
-      expect(asPlainObject(result.compile.compiledGraph.tables)).deep.equals(
-        asPlainObject([
-          {
-            target: {
-              database: "defaultProject",
-              schema: "defaultDataset",
-              name: "action"
-            },
-            canonicalTarget: {
-              database: "defaultProject",
-              schema: "defaultDataset",
-              name: "action"
-            },
-            fileName: "definitions/action.sql",
-            hermeticity: "NON_HERMETIC",
-            query: "SELECT 1",
-            type: "table",
-            enumType: "TABLE",
-            disabled: false
-          }
-        ])
-      );
-    });
-
-    test(`incremental tables can be loaded`, () => {
-      const projectDir = tmpDirFixture.createNewTmpDir();
-      fs.writeFileSync(
-        path.join(projectDir, "workflow_settings.yaml"),
-        VALID_WORKFLOW_SETTINGS_YAML
-      );
-      fs.mkdirSync(path.join(projectDir, "definitions"));
-      fs.writeFileSync(
-        path.join(projectDir, "definitions/actions.yaml"),
-        `
-actions:
-- incrementalTable:
-    filename: action.sql
-    protected: true
-    uniqueKey:
-    -  someKey1
-    -  someKey2`
-      );
-      fs.writeFileSync(path.join(projectDir, "definitions/action.sql"), "SELECT 1");
-
-      const result = runMainInVm(coreExecutionRequestFromPath(projectDir));
-
-      expect(result.compile.compiledGraph.graphErrors.compilationErrors).deep.equals([]);
-      expect(asPlainObject(result.compile.compiledGraph.tables)).deep.equals(
-        asPlainObject([
-          {
-            target: {
-              database: "defaultProject",
-              schema: "defaultDataset",
-              name: "action"
-            },
-            canonicalTarget: {
-              database: "defaultProject",
-              schema: "defaultDataset",
-              name: "action"
-            },
-            fileName: "definitions/action.sql",
-            hermeticity: "NON_HERMETIC",
-            onSchemaChange: "IGNORE",
-            query: "SELECT 1",
-            incrementalQuery: "SELECT 1",
-            type: "incremental",
-            enumType: "INCREMENTAL",
-            protected: true,
-            disabled: false,
-            uniqueKey: ["someKey1", "someKey2"]
-          }
-        ])
-      );
-    });
-
-    test(`views can be loaded`, () => {
-      const projectDir = tmpDirFixture.createNewTmpDir();
-      fs.writeFileSync(
-        path.join(projectDir, "workflow_settings.yaml"),
-        VALID_WORKFLOW_SETTINGS_YAML
-      );
-      fs.mkdirSync(path.join(projectDir, "definitions"));
-      fs.writeFileSync(
-        path.join(projectDir, "definitions/actions.yaml"),
-        `
-actions:
-- view:
-    filename: action.sql`
-      );
-      fs.writeFileSync(path.join(projectDir, "definitions/action.sql"), "SELECT 1");
-
-      const result = runMainInVm(coreExecutionRequestFromPath(projectDir));
-
-      expect(result.compile.compiledGraph.graphErrors.compilationErrors).deep.equals([]);
-      expect(asPlainObject(result.compile.compiledGraph.tables)).deep.equals(
-        asPlainObject([
-          {
-            target: {
-              database: "defaultProject",
-              schema: "defaultDataset",
-              name: "action"
-            },
-            canonicalTarget: {
-              database: "defaultProject",
-              schema: "defaultDataset",
-              name: "action"
-            },
-            fileName: "definitions/action.sql",
-            hermeticity: "NON_HERMETIC",
-            query: "SELECT 1",
-            type: "view",
-            enumType: "VIEW",
-            disabled: false
-          }
-        ])
-      );
-    });
-
-    test(`assertions can be loaded`, () => {
-      const projectDir = tmpDirFixture.createNewTmpDir();
-      fs.writeFileSync(
-        path.join(projectDir, "workflow_settings.yaml"),
-        VALID_WORKFLOW_SETTINGS_YAML
-      );
-      fs.mkdirSync(path.join(projectDir, "definitions"));
-      fs.writeFileSync(
-        path.join(projectDir, "definitions/actions.yaml"),
-        // If change, then change test "sqlx config options checks for assertions".
-        `
-actions:
-- assertion:
-    name: name
-    dataset: dataset
-    project: project
-    dependencyTargets:
-      - name: operation
-        dataset: defaultDataset
-        project: defaultProject
-    filename: action.sql
-    tags:
-      - tagA
-      - tagB
-    disabled: true,
-    description: description
-    hermetic: true,
-    dependOnDependencyAssertions: true`
-      );
-      fs.writeFileSync(path.join(projectDir, "definitions/action.sql"), "SELECT 1");
-      fs.writeFileSync(path.join(projectDir, "definitions/operation.sqlx"), "SELECT 1");
-
-      const result = runMainInVm(coreExecutionRequestFromPath(projectDir));
-
-      expect(result.compile.compiledGraph.graphErrors.compilationErrors).deep.equals([]);
-      expect(asPlainObject(result.compile.compiledGraph.assertions)).deep.equals(
-        asPlainObject([
-          {
-            target: {
-              database: "project",
-              schema: "dataset",
-              name: "name"
-            },
-            canonicalTarget: {
-              database: "project",
-              schema: "dataset",
-              name: "name"
-            },
-            actionDescriptor: {
-              description: "description"
-            },
-            disabled: true,
-            fileName: "definitions/action.sql",
-            hermeticity: "HERMETIC",
-            tags: ["tagA", "tagB"],
-            query: "SELECT 1",
-            dependencyTargets: [
-              {
-                name: "operation",
-                schema: "defaultDataset",
-                database: "defaultProject"
-              }
-            ]
-          }
-        ])
-      );
-    });
-
     test(`fails when file is not found`, () => {
       const projectDir = tmpDirFixture.createNewTmpDir();
       fs.writeFileSync(
@@ -2378,1981 +1030,16 @@ actions:
     });
   });
 
-  suite("sqlx and JS API config options", () => {
-    const exampleActionDescriptor = {
-      inputSqlxConfigBlock: `
-  columns: {
-    column1Key: "column1Val",
-    column2Key: {
-      description: "description",
-      columns: {
-        nestedColumnKey: "nestedColumnVal"
-      },
-      tags: ["tag3", "tag4"],
-      bigqueryPolicyTags: ["bigqueryPolicyTag1", "bigqueryPolicyTag2"],
-    }
-  },`,
-      outputActionDescriptor: {
-        columns: [
-          {
-            description: "column1Val",
-            path: ["column1Key"]
-          },
-          {
-            bigqueryPolicyTags: ["bigqueryPolicyTag1", "bigqueryPolicyTag2"],
-            description: "description",
-            path: ["column2Key"],
-            tags: ["tag3", "tag4"]
-          },
-          {
-            description: "nestedColumnVal",
-            path: ["column2Key", "nestedColumnKey"]
-          }
-        ],
-        description: "description"
-      } as dataform.IColumnDescriptor
-    };
-
-    const exampleBuiltInAssertions = {
-      inputAssertionBlock: `assertions: {
-    uniqueKeys: [["uniqueKey1", "uniqueKey2"]],
-    nonNull: "nonNull",
-    rowConditions: ["rowConditions1", "rowConditions2"],
-  },`,
-      outputAssertions: (filename: string) =>
-        [
-          {
-            target: {
-              database: "defaultProject",
-              schema: "defaultDataset",
-              name: "dataset_name_assertions_uniqueKey_0"
-            },
-            canonicalTarget: {
-              database: "defaultProject",
-              schema: "defaultDataset",
-              name: "dataset_name_assertions_uniqueKey_0"
-            },
-            dependencyTargets: [
-              {
-                database: "project",
-                schema: "dataset",
-                name: "name"
-              }
-            ],
-            disabled: true,
-            fileName: `definitions/${filename}`,
-            parentAction: {
-              database: "project",
-              schema: "dataset",
-              name: "name"
-            },
-            query:
-              "\nSELECT\n  *\nFROM (\n  SELECT\n    uniqueKey1, uniqueKey2,\n    COUNT(1) AS index_row_count\n  FROM `project.dataset.name`\n  GROUP BY uniqueKey1, uniqueKey2\n  ) AS data\nWHERE index_row_count > 1\n",
-            tags: ["tag1", "tag2"]
-          },
-          {
-            target: {
-              database: "defaultProject",
-              schema: "defaultDataset",
-              name: "dataset_name_assertions_rowConditions"
-            },
-            canonicalTarget: {
-              database: "defaultProject",
-              schema: "defaultDataset",
-              name: "dataset_name_assertions_rowConditions"
-            },
-            dependencyTargets: [
-              {
-                database: "project",
-                schema: "dataset",
-                name: "name"
-              }
-            ],
-            disabled: true,
-            fileName: `definitions/${filename}`,
-            parentAction: {
-              database: "project",
-              schema: "dataset",
-              name: "name"
-            },
-            query:
-              "\nSELECT\n  'rowConditions1' AS failing_row_condition,\n  *\nFROM `project.dataset.name`\nWHERE NOT (rowConditions1)\nUNION ALL\nSELECT\n  'rowConditions2' AS failing_row_condition,\n  *\nFROM `project.dataset.name`\nWHERE NOT (rowConditions2)\nUNION ALL\nSELECT\n  'nonNull IS NOT NULL' AS failing_row_condition,\n  *\nFROM `project.dataset.name`\nWHERE NOT (nonNull IS NOT NULL)\n",
-            tags: ["tag1", "tag2"]
-          }
-        ] as dataform.IAssertion[]
-    };
-
-    // If change, then change test "action configs assertions can be loaded".
-    const assertionConfig = `{
-  type: "assertion",
-  name: "name",
-  schema: "dataset",
-  database: "project",
-  dependencies: ["operation"],
-  tags: ["tagA", "tagB"],
-  disabled: true,
-  description: "description",
-  hermetic: true,
-  dependOnDependencyAssertions: true,
-}`;
-    [
-      {
-        filename: "assertion.sqlx",
-        fileContents: `
-config ${assertionConfig}
-SELECT 1`
-      },
-      {
-        filename: "assertion.js",
-        fileContents: `assert("name", ${assertionConfig}).query(ctx => \`\n\nSELECT 1\`)`
-      }
-    ].forEach(testParameters => {
-      test(`for assertions configured in a ${testParameters.filename} file`, () => {
-        const projectDir = tmpDirFixture.createNewTmpDir();
-        fs.writeFileSync(
-          path.join(projectDir, "workflow_settings.yaml"),
-          VALID_WORKFLOW_SETTINGS_YAML
-        );
-        fs.mkdirSync(path.join(projectDir, "definitions"));
-        fs.writeFileSync(path.join(projectDir, "definitions/operation.sqlx"), "SELECT 1");
-        fs.writeFileSync(
-          path.join(projectDir, `definitions/${testParameters.filename}`),
-          testParameters.fileContents
-        );
-
-        const result = runMainInVm(coreExecutionRequestFromPath(projectDir));
-
-        expect(result.compile.compiledGraph.graphErrors.compilationErrors).deep.equals([]);
-        expect(asPlainObject(result.compile.compiledGraph.assertions)).deep.equals(
-          asPlainObject([
-            {
-              target: {
-                database: "project",
-                schema: "dataset",
-                name: "name"
-              },
-              canonicalTarget: {
-                database: "project",
-                schema: "dataset",
-                name: "name"
-              },
-              actionDescriptor: {
-                description: "description"
-              },
-              dependencyTargets: [
-                {
-                  database: "defaultProject",
-                  schema: "defaultDataset",
-                  name: "operation"
-                }
-              ],
-              disabled: true,
-              fileName: `definitions/${testParameters.filename}`,
-              hermeticity: "HERMETIC",
-              tags: ["tagA", "tagB"],
-              query: "\n\nSELECT 1"
-            }
-          ])
-        );
-      });
-    });
-
-    const declarationConfig = `{
-  type: "declaration",
-  name: "name",
-  schema: "dataset",
-  database: "project",
-  description: "description",
-${exampleActionDescriptor.inputSqlxConfigBlock}
-}`;
-
-    [
-      {
-        filename: "declaration.sqlx",
-        fileContents: `
-config ${declarationConfig}`
-      },
-      {
-        filename: "declaration.js",
-        fileContents: `declare(${declarationConfig})`
-      }
-    ].forEach(testParameters => {
-      test(`for declarations configured in a ${testParameters.filename} file`, () => {
-        const projectDir = tmpDirFixture.createNewTmpDir();
-        fs.writeFileSync(
-          path.join(projectDir, "workflow_settings.yaml"),
-          VALID_WORKFLOW_SETTINGS_YAML
-        );
-        fs.mkdirSync(path.join(projectDir, "definitions"));
-        fs.writeFileSync(
-          path.join(projectDir, `definitions/${testParameters.filename}`),
-          testParameters.fileContents
-        );
-
-        const result = runMainInVm(coreExecutionRequestFromPath(projectDir));
-
-        expect(result.compile.compiledGraph.graphErrors.compilationErrors).deep.equals([]);
-        expect(asPlainObject(result.compile.compiledGraph.declarations)).deep.equals(
-          asPlainObject([
-            {
-              target: {
-                database: "project",
-                schema: "dataset",
-                name: "name"
-              },
-              canonicalTarget: {
-                database: "project",
-                schema: "dataset",
-                name: "name"
-              },
-              fileName: `definitions/${testParameters.filename}`,
-              actionDescriptor: exampleActionDescriptor.outputActionDescriptor
-            }
-          ])
-        );
-      });
-    });
-
-    const tableConfig = `{
-  type: "table",
-  name: "name",
-  schema: "dataset",
-  database: "project",
-  dependencies: ["operation"],
-  tags: ["tag1", "tag2"],
-  disabled: true,
-  description: "description",
-${exampleActionDescriptor.inputSqlxConfigBlock}
-  bigquery: {
-    partitionBy: "partitionBy",
-    partitionExpirationDays: 1,
-    requirePartitionFilter: true,
-    clusterBy: ["clusterBy"],
-    labels: {"key": "val"},
-    additionalOptions: {
-      option1Key: "option1",
-      option2Key: "option2",
-    }
-  },
-  ${exampleBuiltInAssertions.inputAssertionBlock}
-  dependOnDependencyAssertions: true,
-  hermetic: true
-}`;
-
-    [
-      {
-        filename: "table.sqlx",
-        fileContents: `
-config ${tableConfig}
-SELECT 1`
-      },
-      {
-        filename: "table.js",
-        fileContents: `publish("name", ${tableConfig}).query(ctx => \`\n\nSELECT 1\`)`
-      }
-    ].forEach(testParameters => {
-      test(`for tables configured in a ${testParameters.filename} file`, () => {
-        const projectDir = tmpDirFixture.createNewTmpDir();
-        fs.writeFileSync(
-          path.join(projectDir, "workflow_settings.yaml"),
-          VALID_WORKFLOW_SETTINGS_YAML
-        );
-        fs.mkdirSync(path.join(projectDir, "definitions"));
-        fs.writeFileSync(path.join(projectDir, "definitions/operation.sqlx"), "SELECT 1");
-        fs.writeFileSync(
-          path.join(projectDir, `definitions/${testParameters.filename}`),
-          testParameters.fileContents
-        );
-
-        const result = runMainInVm(coreExecutionRequestFromPath(projectDir));
-
-        expect(result.compile.compiledGraph.graphErrors.compilationErrors).deep.equals([]);
-        expect(asPlainObject(result.compile.compiledGraph.tables)).deep.equals([
-          {
-            target: {
-              database: "project",
-              schema: "dataset",
-              name: "name"
-            },
-            canonicalTarget: {
-              database: "project",
-              schema: "dataset",
-              name: "name"
-            },
-            type: "table",
-            disabled: true,
-            hermeticity: "HERMETIC",
-            bigquery: {
-              additionalOptions: {
-                option1Key: "option1",
-                option2Key: "option2"
-              },
-              clusterBy: ["clusterBy"],
-              labels: {
-                key: "val"
-              },
-              partitionBy: "partitionBy",
-              partitionExpirationDays: 1,
-              requirePartitionFilter: true
-            },
-            tags: ["tag1", "tag2"],
-            dependencyTargets: [
-              {
-                database: "defaultProject",
-                schema: "defaultDataset",
-                name: "operation"
-              }
-            ],
-            enumType: "TABLE",
-            fileName: `definitions/${testParameters.filename}`,
-            query: "\n\nSELECT 1",
-            actionDescriptor: {
-              ...exampleActionDescriptor.outputActionDescriptor,
-              // sqlxConfig.bigquery.labels are placed as bigqueryLabels.
-              bigqueryLabels: {
-                key: "val"
-              }
-            }
-          }
-        ]);
-        expect(asPlainObject(result.compile.compiledGraph.assertions)).deep.equals(
-          exampleBuiltInAssertions.outputAssertions(testParameters.filename)
-        );
-      });
-    });
-
-    const viewConfig = `{
-  type: "view",
-  name: "name",
-  schema: "dataset",
-  database: "project",
-  dependencies: ["operation"],
-  tags: ["tag1", "tag2"],
-  disabled: true,
-  materialized: true,
-  description: "description",
-${exampleActionDescriptor.inputSqlxConfigBlock}
-  bigquery: {
-    labels: {"key": "val"},
-    additionalOptions: {
-      option1Key: "option1",
-      option2Key: "option2",
-    }
-  },
-  dependOnDependencyAssertions: true,
-  hermetic: true,
-  ${exampleBuiltInAssertions.inputAssertionBlock}
-}`;
-    [
-      {
-        filename: "view.sqlx",
-        fileContents: `
-config ${viewConfig}
-SELECT 1`
-      },
-      {
-        filename: "view.js",
-        fileContents: `publish("name", ${viewConfig}).query(ctx => \`\n\nSELECT 1\`)`
-      }
-    ].forEach(testParameters => {
-      test(`for views configured in a ${testParameters.filename} file`, () => {
-        const projectDir = tmpDirFixture.createNewTmpDir();
-        fs.writeFileSync(
-          path.join(projectDir, "workflow_settings.yaml"),
-          VALID_WORKFLOW_SETTINGS_YAML
-        );
-        fs.mkdirSync(path.join(projectDir, "definitions"));
-        fs.writeFileSync(path.join(projectDir, "definitions/operation.sqlx"), "SELECT 1");
-        fs.writeFileSync(
-          path.join(projectDir, `definitions/${testParameters.filename}`),
-          testParameters.fileContents
-        );
-
-        const result = runMainInVm(coreExecutionRequestFromPath(projectDir));
-
-        expect(result.compile.compiledGraph.graphErrors.compilationErrors).deep.equals([]);
-        expect(asPlainObject(result.compile.compiledGraph.tables)).deep.equals([
-          {
-            target: {
-              database: "project",
-              schema: "dataset",
-              name: "name"
-            },
-            canonicalTarget: {
-              database: "project",
-              schema: "dataset",
-              name: "name"
-            },
-            type: "view",
-            disabled: true,
-            hermeticity: "HERMETIC",
-            bigquery: {
-              additionalOptions: {
-                option1Key: "option1",
-                option2Key: "option2"
-              },
-              labels: {
-                key: "val"
-              }
-            },
-            tags: ["tag1", "tag2"],
-            dependencyTargets: [
-              {
-                database: "defaultProject",
-                schema: "defaultDataset",
-                name: "operation"
-              }
-            ],
-            enumType: "VIEW",
-            fileName: `definitions/${testParameters.filename}`,
-            query: "\n\nSELECT 1",
-            actionDescriptor: {
-              ...exampleActionDescriptor.outputActionDescriptor,
-              // sqlxConfig.bigquery.labels are placed as bigqueryLabels.
-              bigqueryLabels: {
-                key: "val"
-              }
-            },
-            materialized: true
-          }
-        ]);
-        expect(asPlainObject(result.compile.compiledGraph.assertions)).deep.equals(
-          exampleBuiltInAssertions.outputAssertions(testParameters.filename)
-        );
-      });
-    });
-
-    const incrementalTableConfig = `{
-  type: "incremental",
-  name: "name",
-  schema: "dataset",
-  database: "project",
-  dependencies: ["operation"],
-  tags: ["tag1", "tag2"],
-  disabled: true,
-  protected: false,
-  uniqueKey: ["key1", "key2"],
-  description: "description",
-  ${exampleActionDescriptor.inputSqlxConfigBlock}
-  bigquery: {
-    partitionBy: "partitionBy",
-    partitionExpirationDays: 1,
-    requirePartitionFilter: true,
-    updatePartitionFilter: "updatePartitionFilter",
-    clusterBy: ["clusterBy"],
-    labels: {"key": "val"},
-    additionalOptions: {
-      option1Key: "option1",
-      option2Key: "option2",
-    }
-  },
-  dependOnDependencyAssertions: true,
-  ${exampleBuiltInAssertions.inputAssertionBlock}
-  hermetic: true,
-  onSchemaChange: "SYNCHRONIZE",
-}
-`;
-    [
-      {
-        filename: "incremental.sqlx",
-        fileContents: `
-config ${incrementalTableConfig}
-SELECT 1`
-      },
-      {
-        filename: "incremental.js",
-        fileContents: `publish("name", ${incrementalTableConfig}).query(ctx => \`\n\n\nSELECT 1\`)`
-      }
-    ].forEach(testParameters => {
-      test(`for incremental tables configured in a ${testParameters.filename} file`, () => {
-        const projectDir = tmpDirFixture.createNewTmpDir();
-        fs.writeFileSync(
-          path.join(projectDir, "workflow_settings.yaml"),
-          VALID_WORKFLOW_SETTINGS_YAML
-        );
-        fs.mkdirSync(path.join(projectDir, "definitions"));
-        fs.writeFileSync(path.join(projectDir, "definitions/operation.sqlx"), "SELECT 1");
-        fs.writeFileSync(
-          path.join(projectDir, `definitions/${testParameters.filename}`),
-          testParameters.fileContents
-        );
-
-        const result = runMainInVm(coreExecutionRequestFromPath(projectDir));
-
-        expect(result.compile.compiledGraph.graphErrors.compilationErrors).deep.equals([]);
-        expect(asPlainObject(result.compile.compiledGraph.tables)).deep.equals([
-          {
-            target: {
-              database: "project",
-              schema: "dataset",
-              name: "name"
-            },
-            canonicalTarget: {
-              database: "project",
-              schema: "dataset",
-              name: "name"
-            },
-            type: "incremental",
-            disabled: true,
-            protected: false,
-            hermeticity: "HERMETIC",
-            onSchemaChange: "SYNCHRONIZE",
-            bigquery: {
-              additionalOptions: {
-                option1Key: "option1",
-                option2Key: "option2"
-              },
-              clusterBy: ["clusterBy"],
-              labels: {
-                key: "val"
-              },
-              partitionBy: "partitionBy",
-              partitionExpirationDays: 1,
-              requirePartitionFilter: true,
-              updatePartitionFilter: "updatePartitionFilter"
-            },
-            tags: ["tag1", "tag2"],
-            uniqueKey: ["key1", "key2"],
-            dependencyTargets: [
-              {
-                database: "defaultProject",
-                schema: "defaultDataset",
-                name: "operation"
-              }
-            ],
-            enumType: "INCREMENTAL",
-            fileName: `definitions/${testParameters.filename}`,
-            query: "\n\n\nSELECT 1",
-            incrementalQuery: "\n\n\nSELECT 1",
-            actionDescriptor: {
-              ...exampleActionDescriptor.outputActionDescriptor,
-              // sqlxConfig.bigquery.labels are placed as bigqueryLabels.
-              bigqueryLabels: {
-                key: "val"
-              }
-            }
-          }
-        ]);
-        expect(asPlainObject(result.compile.compiledGraph.assertions)).deep.equals(
-          exampleBuiltInAssertions.outputAssertions(testParameters.filename)
-        );
-      });
-    });
-
-    const operationConfig = `{
-  type: "operations",
-  name: "name",
-  schema: "dataset",
-  database: "project",
-  dependencies: ["table"],
-  tags: ["tagA", "tagB"],
-  disabled: true,
-  description: "description",
-  hermetic: true,
-  hasOutput: true,
-  dependOnDependencyAssertions: true,
-${exampleActionDescriptor.inputSqlxConfigBlock}
-}`;
-
-    [
-      {
-        filename: "operation.sqlx",
-        fileContents: `
-config ${operationConfig}
-SELECT 1`
-      },
-      {
-        filename: "operation.js",
-        fileContents: `operate("name", ${operationConfig}).queries(ctx => \`\n\nSELECT 1\`)`
-      }
-    ].forEach(testParameters => {
-      test(`for operations configured in a ${testParameters.filename} file`, () => {
-        const projectDir = tmpDirFixture.createNewTmpDir();
-        fs.writeFileSync(
-          path.join(projectDir, "workflow_settings.yaml"),
-          VALID_WORKFLOW_SETTINGS_YAML
-        );
-        fs.mkdirSync(path.join(projectDir, "definitions"));
-        fs.writeFileSync(
-          path.join(projectDir, "definitions/table.sqlx"),
-          `config {type: "view"} SELECT 1`
-        );
-        fs.writeFileSync(
-          path.join(projectDir, `definitions/${testParameters.filename}`),
-          testParameters.fileContents
-        );
-
-        const result = runMainInVm(coreExecutionRequestFromPath(projectDir));
-
-        expect(result.compile.compiledGraph.graphErrors.compilationErrors).deep.equals([]);
-        expect(asPlainObject(result.compile.compiledGraph.operations)).deep.equals(
-          asPlainObject([
-            {
-              target: {
-                database: "project",
-                schema: "dataset",
-                name: "name"
-              },
-              canonicalTarget: {
-                database: "project",
-                schema: "dataset",
-                name: "name"
-              },
-              dependencyTargets: [
-                {
-                  database: "defaultProject",
-                  schema: "defaultDataset",
-                  name: "table"
-                }
-              ],
-              disabled: true,
-              fileName: `definitions/${testParameters.filename}`,
-              hermeticity: "HERMETIC",
-              hasOutput: true,
-              tags: ["tagA", "tagB"],
-              queries: ["\n\nSELECT 1"],
-              actionDescriptor: exampleActionDescriptor.outputActionDescriptor
-            }
-          ])
-        );
-      });
-    });
-
-    ["table", "view", "incremental"].forEach(tableType => {
-      [`"fieldValue"`, `["fieldValue"]`].forEach(uniqueKeyField => {
-        test(`for ${tableType} built-in assertions uniqueKey with value ${uniqueKeyField}`, () => {
-          // The `uniqueKey` built in assertion field cannot be present at the same time as
-          // `uniqueKeys`, so it is tested separately here.
-          const projectDir = tmpDirFixture.createNewTmpDir();
-          fs.writeFileSync(
-            path.join(projectDir, "workflow_settings.yaml"),
-            VALID_WORKFLOW_SETTINGS_YAML
-          );
-          fs.mkdirSync(path.join(projectDir, "definitions"));
-          fs.writeFileSync(
-            path.join(projectDir, "definitions/filename.sqlx"),
-            `
-config {
-  type: "${tableType}",
-  assertions: {
-    uniqueKey: ${uniqueKeyField},
-  },
-}
-SELECT 2`
-          );
-
-          const result = runMainInVm(coreExecutionRequestFromPath(projectDir));
-
-          expect(result.compile.compiledGraph.graphErrors.compilationErrors).deep.equals([]);
-          expect(asPlainObject(result.compile.compiledGraph.assertions)).deep.equals([
-            {
-              target: {
-                database: "defaultProject",
-                schema: "defaultDataset",
-                name: "defaultDataset_filename_assertions_uniqueKey_0"
-              },
-              canonicalTarget: {
-                database: "defaultProject",
-                schema: "defaultDataset",
-                name: "defaultDataset_filename_assertions_uniqueKey_0"
-              },
-              dependencyTargets: [
-                {
-                  database: "defaultProject",
-                  schema: "defaultDataset",
-                  name: "filename"
-                }
-              ],
-              fileName: "definitions/filename.sqlx",
-              parentAction: {
-                database: "defaultProject",
-                schema: "defaultDataset",
-                name: "filename"
-              },
-              query:
-                "\nSELECT\n  *\nFROM (\n  SELECT\n    fieldValue,\n    COUNT(1) AS index_row_count\n  FROM `defaultProject.defaultDataset.filename`\n  GROUP BY fieldValue\n  ) AS data\nWHERE index_row_count > 1\n"
-            }
-          ]);
-        });
-      });
-    });
-
-    test(`for notebooks`, () => {
-      const projectDir = tmpDirFixture.createNewTmpDir();
-      fs.writeFileSync(
-        path.join(projectDir, "workflow_settings.yaml"),
-        VALID_WORKFLOW_SETTINGS_YAML
-      );
-      fs.mkdirSync(path.join(projectDir, "definitions"));
-      fs.writeFileSync(path.join(projectDir, "definitions/operation.sqlx"), "SELECT 1");
-      fs.writeFileSync(
-        path.join(projectDir, "definitions/filename.ipynb"),
-        EMPTY_NOTEBOOK_CONTENTS
-      );
-      fs.writeFileSync(
-        path.join(projectDir, "definitions/notebook.js"),
-        `
-notebook({
-  name: "name",
-  location: "location",
-  project: "project",
-  dependencyTargets: [{
-    name: "operation",
-  }],
-  filename: "filename.ipynb",
-  tags: ["tagA", "tagB"],
-  disabled: true,
-  description: "description",
-  dependOnDependencyAssertions: true
-})`
-      );
-
-      const result = runMainInVm(coreExecutionRequestFromPath(projectDir));
-
-      expect(result.compile.compiledGraph.graphErrors.compilationErrors).deep.equals([]);
-      expect(asPlainObject(result.compile.compiledGraph.notebooks)).deep.equals(
-        asPlainObject([
-          {
-            target: {
-              database: "project",
-              schema: "location",
-              name: "name"
-            },
-            canonicalTarget: {
-              database: "project",
-              schema: "location",
-              name: "name"
-            },
-            dependencyTargets: [
-              {
-                database: "defaultProject",
-                schema: "defaultDataset",
-                name: "operation"
-              }
-            ],
-            disabled: true,
-            fileName: "definitions/filename.ipynb",
-            tags: ["tagA", "tagB"],
-            notebookContents: `{"cells":[]}`
-          }
-        ])
-      );
-    });
-  });
-
-  suite("action config options", () => {
-    const exampleBuiltInAssertions = {
-      inputActionConfigBlock: `
-    assertions:
-      uniqueKeys:
-      - uniqueKey:
-        - uniqueKey1
-        - uniqueKey2
-      nonNull:
-      - nonNull
-      rowConditions:
-      - rowConditions1
-      - rowConditions2
-`,
-      outputAssertions: [
-        {
-          target: {
-            database: "defaultProject",
-            schema: "defaultDataset",
-            name: "dataset_name_assertions_uniqueKey_0"
-          },
-          canonicalTarget: {
-            database: "defaultProject",
-            schema: "defaultDataset",
-            name: "dataset_name_assertions_uniqueKey_0"
-          },
-          dependencyTargets: [
-            {
-              database: "project",
-              schema: "dataset",
-              name: "name"
-            }
-          ],
-          disabled: true,
-          // It would make more sense for this to be the path to the config, but we haven't yet
-          // introduced a configFilename field.
-          fileName: "definitions/filename.sql",
-          parentAction: {
-            database: "project",
-            schema: "dataset",
-            name: "name"
-          },
-          query:
-            "\nSELECT\n  *\nFROM (\n  SELECT\n    uniqueKey1, uniqueKey2,\n    COUNT(1) AS index_row_count\n  FROM `project.dataset.name`\n  GROUP BY uniqueKey1, uniqueKey2\n  ) AS data\nWHERE index_row_count > 1\n",
-          tags: ["tag1", "tag2"]
-        },
-        {
-          target: {
-            database: "defaultProject",
-            schema: "defaultDataset",
-            name: "dataset_name_assertions_rowConditions"
-          },
-          canonicalTarget: {
-            database: "defaultProject",
-            schema: "defaultDataset",
-            name: "dataset_name_assertions_rowConditions"
-          },
-          dependencyTargets: [
-            {
-              database: "project",
-              schema: "dataset",
-              name: "name"
-            }
-          ],
-          disabled: true,
-          // It would make more sense for this to be the path to the config, but we haven't yet
-          // introduced a configFilename field.
-          fileName: "definitions/filename.sql",
-          parentAction: {
-            database: "project",
-            schema: "dataset",
-            name: "name"
-          },
-          query:
-            "\nSELECT\n  'rowConditions1' AS failing_row_condition,\n  *\nFROM `project.dataset.name`\nWHERE NOT (rowConditions1)\nUNION ALL\nSELECT\n  'rowConditions2' AS failing_row_condition,\n  *\nFROM `project.dataset.name`\nWHERE NOT (rowConditions2)\nUNION ALL\nSELECT\n  'nonNull IS NOT NULL' AS failing_row_condition,\n  *\nFROM `project.dataset.name`\nWHERE NOT (nonNull IS NOT NULL)\n",
-          tags: ["tag1", "tag2"]
-        }
-      ] as dataform.IAssertion[]
-    };
-
-    test(`for assertions`, () => {
-      const projectDir = tmpDirFixture.createNewTmpDir();
-      fs.writeFileSync(
-        path.join(projectDir, "workflow_settings.yaml"),
-        VALID_WORKFLOW_SETTINGS_YAML
-      );
-      fs.mkdirSync(path.join(projectDir, "definitions"));
-      fs.writeFileSync(path.join(projectDir, "definitions/operation.sqlx"), "SELECT 1");
-      fs.writeFileSync(path.join(projectDir, "definitions/filename.sql"), "SELECT 1");
-      fs.writeFileSync(
-        path.join(projectDir, "definitions/actions.yaml"),
-        `
-actions:
-- assertion:
-    name: name
-    dataset: dataset
-    project: project
-    dependencyTargets:
-    - name: operation
-    filename: filename.sql
-    tags:
-    - tagA
-    - tagB
-    disabled: true
-    description: description
-    hermetic: true
-    dependOnDependencyAssertions: true
-`
-      );
-
-      const result = runMainInVm(coreExecutionRequestFromPath(projectDir));
-
-      expect(result.compile.compiledGraph.graphErrors.compilationErrors).deep.equals([]);
-      expect(asPlainObject(result.compile.compiledGraph.assertions)).deep.equals(
-        asPlainObject([
-          {
-            target: {
-              database: "project",
-              schema: "dataset",
-              name: "name"
-            },
-            canonicalTarget: {
-              database: "project",
-              schema: "dataset",
-              name: "name"
-            },
-            actionDescriptor: {
-              description: "description"
-            },
-            dependencyTargets: [
-              {
-                database: "defaultProject",
-                schema: "defaultDataset",
-                name: "operation"
-              }
-            ],
-            disabled: true,
-            fileName: "definitions/filename.sql",
-            hermeticity: "HERMETIC",
-            tags: ["tagA", "tagB"],
-            query: "SELECT 1"
-          }
-        ])
-      );
-    });
-
-    test(`for declarations`, () => {
-      const projectDir = tmpDirFixture.createNewTmpDir();
-      fs.writeFileSync(
-        path.join(projectDir, "workflow_settings.yaml"),
-        VALID_WORKFLOW_SETTINGS_YAML
-      );
-      fs.mkdirSync(path.join(projectDir, "definitions"));
-      fs.writeFileSync(path.join(projectDir, "definitions/operation.sqlx"), "SELECT 1");
-      fs.writeFileSync(
-        path.join(projectDir, "definitions/actions.yaml"),
-        `
-actions:
-- declaration:
-    name: name
-    dataset: dataset
-    project: project
-    description: description
-`
-      );
-
-      const result = runMainInVm(coreExecutionRequestFromPath(projectDir));
-
-      expect(result.compile.compiledGraph.graphErrors.compilationErrors).deep.equals([]);
-      expect(asPlainObject(result.compile.compiledGraph.declarations)).deep.equals(
-        asPlainObject([
-          {
-            target: {
-              database: "project",
-              schema: "dataset",
-              name: "name"
-            },
-            canonicalTarget: {
-              database: "project",
-              schema: "dataset",
-              name: "name"
-            },
-            actionDescriptor: {
-              description: "description"
-            }
-          }
-        ])
-      );
-    });
-
-    test("for tables", () => {
-      const projectDir = tmpDirFixture.createNewTmpDir();
-      fs.writeFileSync(
-        path.join(projectDir, "workflow_settings.yaml"),
-        VALID_WORKFLOW_SETTINGS_YAML
-      );
-      fs.mkdirSync(path.join(projectDir, "definitions"));
-      fs.writeFileSync(path.join(projectDir, "definitions/operation.sqlx"), "SELECT 1");
-      fs.writeFileSync(path.join(projectDir, "definitions/filename.sql"), "SELECT 1");
-      fs.writeFileSync(
-        path.join(projectDir, "definitions/actions.yaml"),
-        `
-actions:
-- table:
-    name: name
-    dataset: dataset
-    project: project
-    dependencyTargets:
-    - name: operation
-    filename: filename.sql
-    tags:
-    - tag1
-    - tag2
-    disabled: true
-    description: description
-    partitionBy: partitionBy
-    partitionExpirationDays: 1
-    requirePartitionFilter: true
-    clusterBy:
-    - clusterBy
-    labels:
-      key: val
-    additionalOptions:
-      option1Key: option1
-      option2Key: option2
-    dependOnDependencyAssertions: true
-    hermetic: true
-${exampleBuiltInAssertions.inputActionConfigBlock}
-    `
-      );
-
-      const result = runMainInVm(coreExecutionRequestFromPath(projectDir));
-
-      expect(result.compile.compiledGraph.graphErrors.compilationErrors).deep.equals([]);
-      expect(asPlainObject(result.compile.compiledGraph.tables)).deep.equals([
-        {
-          target: {
-            database: "project",
-            schema: "dataset",
-            name: "name"
-          },
-          canonicalTarget: {
-            database: "project",
-            schema: "dataset",
-            name: "name"
-          },
-          type: "table",
-          disabled: true,
-          hermeticity: "HERMETIC",
-          bigquery: {
-            additionalOptions: {
-              option1Key: "option1",
-              option2Key: "option2"
-            },
-            clusterBy: ["clusterBy"],
-            labels: {
-              key: "val"
-            },
-            partitionBy: "partitionBy",
-            partitionExpirationDays: 1,
-            requirePartitionFilter: true
-          },
-          tags: ["tag1", "tag2"],
-          dependencyTargets: [
-            {
-              database: "defaultProject",
-              schema: "defaultDataset",
-              name: "operation"
-            }
-          ],
-          enumType: "TABLE",
-          fileName: "definitions/filename.sql",
-          query: "SELECT 1",
-          actionDescriptor: {
-            bigqueryLabels: {
-              key: "val"
-            },
-            description: "description"
-          }
-        }
-      ]);
-      expect(asPlainObject(result.compile.compiledGraph.assertions)).deep.equals(
-        exampleBuiltInAssertions.outputAssertions
-      );
-    });
-
-    test("for views", () => {
-      const projectDir = tmpDirFixture.createNewTmpDir();
-      fs.writeFileSync(
-        path.join(projectDir, "workflow_settings.yaml"),
-        VALID_WORKFLOW_SETTINGS_YAML
-      );
-      fs.mkdirSync(path.join(projectDir, "definitions"));
-      fs.writeFileSync(path.join(projectDir, "definitions/operation.sqlx"), "SELECT 1");
-      fs.writeFileSync(path.join(projectDir, "definitions/filename.sql"), "SELECT 1");
-      fs.writeFileSync(
-        path.join(projectDir, "definitions/actions.yaml"),
-        `
-actions:
-- view:
-    name: name
-    dataset: dataset
-    project: project
-    dependencyTargets:
-    - name: operation
-    filename: filename.sql
-    tags:
-    - tag1
-    - tag2
-    disabled: true
-    materialized: true
-    description: description
-    labels:
-      key: val
-    additionalOptions:
-      option1Key: option1
-      option2Key: option2
-    dependOnDependencyAssertions: true
-${exampleBuiltInAssertions.inputActionConfigBlock}
-    hermetic: true
-    `
-      );
-
-      const result = runMainInVm(coreExecutionRequestFromPath(projectDir));
-
-      expect(result.compile.compiledGraph.graphErrors.compilationErrors).deep.equals([]);
-      expect(asPlainObject(result.compile.compiledGraph.tables)).deep.equals([
-        {
-          target: {
-            database: "project",
-            schema: "dataset",
-            name: "name"
-          },
-          canonicalTarget: {
-            database: "project",
-            schema: "dataset",
-            name: "name"
-          },
-          type: "view",
-          disabled: true,
-          hermeticity: "HERMETIC",
-          bigquery: {
-            additionalOptions: {
-              option1Key: "option1",
-              option2Key: "option2"
-            },
-            labels: {
-              key: "val"
-            }
-          },
-          tags: ["tag1", "tag2"],
-          dependencyTargets: [
-            {
-              database: "defaultProject",
-              schema: "defaultDataset",
-              name: "operation"
-            }
-          ],
-          enumType: "VIEW",
-          fileName: "definitions/filename.sql",
-          query: "SELECT 1",
-          actionDescriptor: {
-            bigqueryLabels: {
-              key: "val"
-            },
-            description: "description"
-          },
-          materialized: true
-        }
-      ]);
-      expect(asPlainObject(result.compile.compiledGraph.assertions)).deep.equals(
-        exampleBuiltInAssertions.outputAssertions
-      );
-    });
-
-    test("for incremental tables", () => {
-      const projectDir = tmpDirFixture.createNewTmpDir();
-      fs.writeFileSync(
-        path.join(projectDir, "workflow_settings.yaml"),
-        VALID_WORKFLOW_SETTINGS_YAML
-      );
-      fs.mkdirSync(path.join(projectDir, "definitions"));
-      fs.writeFileSync(path.join(projectDir, "definitions/operation.sqlx"), "SELECT 1");
-      fs.writeFileSync(path.join(projectDir, "definitions/filename.sql"), "SELECT 1");
-      fs.writeFileSync(
-        path.join(projectDir, "definitions/actions.yaml"),
-        `
-actions:
-- incrementalTable:
-    name: name
-    dataset: dataset
-    project: project
-    dependencyTargets:
-    - name: operation
-    filename: filename.sql
-    tags:
-    - tag1
-    - tag2
-    disabled: true
-    protected: true
-    uniqueKey:
-    - key1
-    - key2
-    description: description
-    partitionBy: partitionBy
-    partitionExpirationDays: 1
-    requirePartitionFilter: true
-    updatePartitionFilter: "updatePartitionFilter"
-    clusterBy:
-    - clusterBy
-    labels:
-      key: val
-    additionalOptions:
-      option1Key: option1
-      option2Key: option2
-    dependOnDependencyAssertions: true
-${exampleBuiltInAssertions.inputActionConfigBlock}
-    hermetic: true
-    onSchemaChange: FAIL
-    `
-      );
-
-      const result = runMainInVm(coreExecutionRequestFromPath(projectDir));
-
-      expect(result.compile.compiledGraph.graphErrors.compilationErrors).deep.equals([]);
-      expect(asPlainObject(result.compile.compiledGraph.tables)).deep.equals([
-        {
-          target: {
-            database: "project",
-            schema: "dataset",
-            name: "name"
-          },
-          canonicalTarget: {
-            database: "project",
-            schema: "dataset",
-            name: "name"
-          },
-          type: "incremental",
-          disabled: true,
-          protected: true,
-          hermeticity: "HERMETIC",
-          onSchemaChange: "FAIL",
-          bigquery: {
-            additionalOptions: {
-              option1Key: "option1",
-              option2Key: "option2"
-            },
-            clusterBy: ["clusterBy"],
-            labels: {
-              key: "val"
-            },
-            partitionBy: "partitionBy",
-            partitionExpirationDays: 1,
-            requirePartitionFilter: true,
-            updatePartitionFilter: "updatePartitionFilter"
-          },
-          tags: ["tag1", "tag2"],
-          uniqueKey: ["key1", "key2"],
-          dependencyTargets: [
-            {
-              database: "defaultProject",
-              schema: "defaultDataset",
-              name: "operation"
-            }
-          ],
-          enumType: "INCREMENTAL",
-          fileName: "definitions/filename.sql",
-          query: "SELECT 1",
-          incrementalQuery: "SELECT 1",
-          actionDescriptor: {
-            bigqueryLabels: {
-              key: "val"
-            },
-            description: "description"
-          }
-        }
-      ]);
-      expect(asPlainObject(result.compile.compiledGraph.assertions)).deep.equals(
-        exampleBuiltInAssertions.outputAssertions
-      );
-    });
-
-    test(`for operations`, () => {
-      const projectDir = tmpDirFixture.createNewTmpDir();
-      fs.writeFileSync(
-        path.join(projectDir, "workflow_settings.yaml"),
-        VALID_WORKFLOW_SETTINGS_YAML
-      );
-      fs.mkdirSync(path.join(projectDir, "definitions"));
-      fs.writeFileSync(
-        path.join(projectDir, "definitions/table.sqlx"),
-        `config {type: "view"} SELECT 1`
-      );
-      fs.writeFileSync(path.join(projectDir, "definitions/filename.sql"), "SELECT 1");
-      fs.writeFileSync(
-        path.join(projectDir, "definitions/actions.yaml"),
-        `
-actions:
-- operation:
-    name: name
-    dataset: dataset
-    project: project
-    dependencyTargets:
-    - name: table
-    filename: filename.sql
-    tags:
-    - tagA
-    - tagB
-    disabled: true
-    hasOutput: true
-    description: description
-    dependOnDependencyAssertions: true
-    hermetic: true
-    `
-      );
-
-      const result = runMainInVm(coreExecutionRequestFromPath(projectDir));
-
-      expect(result.compile.compiledGraph.graphErrors.compilationErrors).deep.equals([]);
-      expect(asPlainObject(result.compile.compiledGraph.operations)).deep.equals(
-        asPlainObject([
-          {
-            target: {
-              database: "project",
-              schema: "dataset",
-              name: "name"
-            },
-            canonicalTarget: {
-              database: "project",
-              schema: "dataset",
-              name: "name"
-            },
-            dependencyTargets: [
-              {
-                database: "defaultProject",
-                schema: "defaultDataset",
-                name: "table"
-              }
-            ],
-            disabled: true,
-            fileName: "definitions/filename.sql",
-            hermeticity: "HERMETIC",
-            hasOutput: true,
-            tags: ["tagA", "tagB"],
-            queries: ["SELECT 1"],
-            actionDescriptor: {
-              description: "description"
-            }
-          }
-        ])
-      );
-    });
-
-    test(`for notebooks`, () => {
-      const projectDir = tmpDirFixture.createNewTmpDir();
-      fs.writeFileSync(
-        path.join(projectDir, "workflow_settings.yaml"),
-        VALID_WORKFLOW_SETTINGS_YAML
-      );
-      fs.mkdirSync(path.join(projectDir, "definitions"));
-
-      fs.writeFileSync(path.join(projectDir, "definitions/operation.sqlx"), "SELECT 1");
-      fs.writeFileSync(
-        path.join(projectDir, "definitions/filename.ipynb"),
-        EMPTY_NOTEBOOK_CONTENTS
-      );
-      fs.writeFileSync(
-        path.join(projectDir, "definitions/actions.yaml"),
-        `
-actions:
-- notebook:
-    name: name
-    location: location
-    project: project
-    dependencyTargets:
-    - name: operation
-    filename: filename.ipynb
-    tags:
-    - tagA
-    - tagB
-    disabled: true
-    description: description
-    dependOnDependencyAssertions: true`
-      );
-
-      const result = runMainInVm(coreExecutionRequestFromPath(projectDir));
-
-      expect(result.compile.compiledGraph.graphErrors.compilationErrors).deep.equals([]);
-      expect(asPlainObject(result.compile.compiledGraph.notebooks)).deep.equals(
-        asPlainObject([
-          {
-            target: {
-              database: "project",
-              schema: "location",
-              name: "name"
-            },
-            canonicalTarget: {
-              database: "project",
-              schema: "location",
-              name: "name"
-            },
-            dependencyTargets: [
-              {
-                database: "defaultProject",
-                schema: "defaultDataset",
-                name: "operation"
-              }
-            ],
-            disabled: true,
-            fileName: "definitions/filename.ipynb",
-            tags: ["tagA", "tagB"],
-            notebookContents: `{"cells":[]}`
-          }
-        ])
-      );
-    });
-  });
-
-  suite("Assertions as dependencies", ({ beforeEach }) => {
-    [
-      TestConfigs.bigquery,
-      TestConfigs.bigqueryWithDatasetSuffix,
-      TestConfigs.bigqueryWithNamePrefix
-    ].forEach(testConfig => {
-      let projectDir: any;
-      beforeEach("Create temporary dir and files", () => {
-        projectDir = tmpDirFixture.createNewTmpDir();
-        fs.writeFileSync(
-          path.join(projectDir, "workflow_settings.yaml"),
-          dumpYaml(dataform.WorkflowSettings.create(testConfig))
-        );
-        fs.mkdirSync(path.join(projectDir, "definitions"));
-        fs.writeFileSync(
-          path.join(projectDir, "definitions/A.sqlx"),
-          `
-config {
-  type: "table",
-  assertions: {rowConditions: ["test > 1"]}}
-  SELECT 1 as test`
-        );
-        fs.writeFileSync(
-          path.join(projectDir, "definitions/A_assert.sqlx"),
-          `
-config {
-  type: "assertion",
-}
-select test from \${ref("A")} where test > 3`
-        );
-        fs.writeFileSync(path.join(projectDir, "definitions/B.sql"), "SELECT 1");
-        fs.writeFileSync(path.join(projectDir, "definitions/C.sql"), "SELECT 1");
-        fs.writeFileSync(
-          path.join(projectDir, `definitions/notebook.ipynb`),
-          EMPTY_NOTEBOOK_CONTENTS
-        );
-      });
-
-      test("When dependOnDependencyAssertions property is set to true, assertions from A are added as dependencies", () => {
-        fs.writeFileSync(
-          path.join(projectDir, "definitions/B.sqlx"),
-          `
-config {
-  type: "table",
-  dependOnDependencyAssertions: true,
-  dependencies: ["A"]
-}
-select 1 as btest
-`
-        );
-
-        const result = runMainInVm(coreExecutionRequestFromPath(projectDir));
-
-        expect(result.compile.compiledGraph.graphErrors.compilationErrors).deep.equals([]);
-        expect(
-          asPlainObject(
-            result.compile.compiledGraph.tables.find(
-              table => table.target.name === prefixAdjustedName(testConfig.namePrefix, "B")
-            ).dependencyTargets.length
-          )
-        ).equals(3);
-        expect(
-          asPlainObject(
-            result.compile.compiledGraph.tables
-              .find(table => table.target.name === prefixAdjustedName(testConfig.namePrefix, "B"))
-              .dependencyTargets.flatMap(dependencyTarget => dependencyTarget.name)
-          )
-        ).deep.equals([
-          prefixAdjustedName(testConfig.namePrefix, "A"),
-          prefixAdjustedName(testConfig.namePrefix, "defaultDataset_A_assertions_rowConditions"),
-          prefixAdjustedName(testConfig.namePrefix, "A_assert")
-        ]);
-      });
-
-      test("Setting includeDependentAssertions to true in config.dependencies adds assertions from that dependency to dependencyTargets", () => {
-        fs.writeFileSync(
-          path.join(projectDir, "definitions/B.sqlx"),
-          `
-config {
-  type: "table",
-  dependencies: [{name: "A", includeDependentAssertions: true}, "C"]
-}
-select 1 as btest`
-        );
-        fs.writeFileSync(
-          path.join(projectDir, "definitions/C.sqlx"),
-          `
-config {
-  type: "table",
-  assertions: {
-    rowConditions: ["test > 1"]
-  }
-}
-SELECT 1 as test`
-        );
-
-        const result = runMainInVm(coreExecutionRequestFromPath(projectDir));
-
-        expect(result.compile.compiledGraph.graphErrors.compilationErrors).deep.equals([]);
-        expect(
-          asPlainObject(
-            result.compile.compiledGraph.tables.find(
-              table => table.target.name === prefixAdjustedName(testConfig.namePrefix, "B")
-            ).dependencyTargets.length
-          )
-        ).equals(4);
-        expect(
-          asPlainObject(
-            result.compile.compiledGraph.tables
-              .find(table => table.target.name === prefixAdjustedName(testConfig.namePrefix, "B"))
-              .dependencyTargets.flatMap(dependencyTarget => dependencyTarget.name)
-          )
-        ).deep.equals([
-          prefixAdjustedName(testConfig.namePrefix, "A"),
-          prefixAdjustedName(testConfig.namePrefix, "defaultDataset_A_assertions_rowConditions"),
-          prefixAdjustedName(testConfig.namePrefix, "A_assert"),
-          prefixAdjustedName(testConfig.namePrefix, "C")
-        ]);
-      });
-
-      test("Setting includeDependentAssertions to true in ref, adds assertions from that dependency to dependencyTargets", () => {
-        fs.writeFileSync(
-          path.join(projectDir, "definitions/B.sqlx"),
-          `
-config {
-  type: "table",
-  dependencies: ["A"]
-}
-select * from \${ref({name: "C", includeDependentAssertions: true})}
-select 1 as btest`
-        );
-        fs.writeFileSync(
-          path.join(projectDir, "definitions/C.sqlx"),
-          `
-config {
-  type: "table",
-    assertions: {
-      rowConditions: ["test > 1"]
-  }
-}
-SELECT 1 as test`
-        );
-
-        const result = runMainInVm(coreExecutionRequestFromPath(projectDir));
-
-        expect(result.compile.compiledGraph.graphErrors.compilationErrors).deep.equals([]);
-        expect(
-          asPlainObject(
-            result.compile.compiledGraph.tables.find(
-              table => table.target.name === prefixAdjustedName(testConfig.namePrefix, "B")
-            ).dependencyTargets.length
-          )
-        ).equals(3);
-        expect(
-          asPlainObject(
-            result.compile.compiledGraph.tables
-              .find(table => table.target.name === prefixAdjustedName(testConfig.namePrefix, "B"))
-              .dependencyTargets.flatMap(dependencyTarget => dependencyTarget.name)
-          )
-        ).deep.equals([
-          prefixAdjustedName(testConfig.namePrefix, "A"),
-          prefixAdjustedName(testConfig.namePrefix, "C"),
-          prefixAdjustedName(testConfig.namePrefix, "defaultDataset_C_assertions_rowConditions")
-        ]);
-      });
-
-      test("When dependOnDependencyAssertions=true and includeDependentAssertions=false, the assertions related to dependency should not be added to dependencyTargets", () => {
-        fs.writeFileSync(
-          path.join(projectDir, "definitions/B.sqlx"),
-          `
-config {
-  type: "table",
-  dependOnDependencyAssertions: true,
-  dependencies: ["A"]
-}
-select * from \${ref({name: "C", includeDependentAssertions: false})}
-select 1 as btest`
-        );
-        fs.writeFileSync(
-          path.join(projectDir, "definitions/C.sqlx"),
-          `
-config {
-  type: "table",
-    assertions: {
-      rowConditions: ["test > 1"]
-  }
-}
-SELECT 1 as test`
-        );
-
-        const result = runMainInVm(coreExecutionRequestFromPath(projectDir));
-
-        expect(result.compile.compiledGraph.graphErrors.compilationErrors).deep.equals([]);
-        expect(
-          asPlainObject(
-            result.compile.compiledGraph.tables.find(
-              table => table.target.name === prefixAdjustedName(testConfig.namePrefix, "B")
-            ).dependencyTargets.length
-          )
-        ).equals(4);
-        expect(
-          asPlainObject(
-            result.compile.compiledGraph.tables
-              .find(table => table.target.name === prefixAdjustedName(testConfig.namePrefix, "B"))
-              .dependencyTargets.flatMap(dependencyTarget => dependencyTarget.name)
-          )
-        ).deep.equals([
-          prefixAdjustedName(testConfig.namePrefix, "A"),
-          prefixAdjustedName(testConfig.namePrefix, "defaultDataset_A_assertions_rowConditions"),
-          prefixAdjustedName(testConfig.namePrefix, "A_assert"),
-          prefixAdjustedName(testConfig.namePrefix, "C")
-        ]);
-      });
-
-      test("When dependOnDependencyAssertions=false and includeDependentAssertions=true, the assertions related to dependency should be added to dependencyTargets", () => {
-        fs.writeFileSync(
-          path.join(projectDir, "definitions/B.sqlx"),
-          `
-config {
-  type: "operations",
-  dependOnDependencyAssertions: false,
-  dependencies: ["A"]
-}
-select * from \${ref({name: "C", includeDependentAssertions: true})}
-select 1 as btest`
-        );
-        fs.writeFileSync(
-          path.join(projectDir, "definitions/C.sqlx"),
-          `
-config {
-  type: "table",
-    assertions: {
-      rowConditions: ["test > 1"]
-  }
-}
-SELECT 1 as test`
-        );
-
-        const result = runMainInVm(coreExecutionRequestFromPath(projectDir));
-
-        expect(result.compile.compiledGraph.graphErrors.compilationErrors).deep.equals([]);
-        expect(
-          asPlainObject(
-            result.compile.compiledGraph.operations.find(
-              operation => operation.target.name === prefixAdjustedName(testConfig.namePrefix, "B")
-            ).dependencyTargets.length
-          )
-        ).equals(3);
-        expect(
-          asPlainObject(
-            result.compile.compiledGraph.operations
-              .find(
-                operation =>
-                  operation.target.name === prefixAdjustedName(testConfig.namePrefix, "B")
-              )
-              .dependencyTargets.flatMap(dependencyTarget => dependencyTarget.name)
-          )
-        ).deep.equals([
-          prefixAdjustedName(testConfig.namePrefix, "A"),
-          prefixAdjustedName(testConfig.namePrefix, "C"),
-          prefixAdjustedName(testConfig.namePrefix, "defaultDataset_C_assertions_rowConditions")
-        ]);
-      });
-
-      test("Assertions added through includeDependentAssertions and explicitly listed in dependencies are deduplicated.", () => {
-        fs.writeFileSync(
-          path.join(projectDir, "definitions/B.sqlx"),
-          `
-config {
-  type: "table",
-  dependencies: ["A_assert"]
-}
-select * from \${ref({name: "A", includeDependentAssertions: true})}
-select 1 as btest`
-        );
-
-        const result = runMainInVm(coreExecutionRequestFromPath(projectDir));
-
-        expect(result.compile.compiledGraph.graphErrors.compilationErrors).deep.equals([]);
-        expect(
-          asPlainObject(
-            result.compile.compiledGraph.tables.find(
-              table => table.target.name === prefixAdjustedName(testConfig.namePrefix, "B")
-            ).dependencyTargets.length
-          )
-        ).equals(3);
-        expect(
-          asPlainObject(
-            result.compile.compiledGraph.tables
-              .find(table => table.target.name === prefixAdjustedName(testConfig.namePrefix, "B"))
-              .dependencyTargets.flatMap(dependencyTarget => dependencyTarget.name)
-          )
-        ).deep.equals([
-          prefixAdjustedName(testConfig.namePrefix, "A_assert"),
-          prefixAdjustedName(testConfig.namePrefix, "A"),
-          prefixAdjustedName(testConfig.namePrefix, "defaultDataset_A_assertions_rowConditions")
-        ]);
-      });
-
-      test("When includeDependentAssertions property in config and ref are set differently for the same dependency, compilation error is thrown.", () => {
-        fs.writeFileSync(
-          path.join(projectDir, "definitions/B.sqlx"),
-          `
-config {
-  type: "table",
-  dependencies: [{name: "A", includeDependentAssertions: false}, {name: "C", includeDependentAssertions: true}]
-}
-select * from \${ref({name: "A", includeDependentAssertions: true})}
-select * from \${ref({name: "C", includeDependentAssertions: false})}
-select 1 as btest`
-        );
-        fs.writeFileSync(
-          path.join(projectDir, "definitions/C.sqlx"),
-          `
-config {
-  type: "table",
-    assertions: {
-      rowConditions: ["test > 1"]
-  }
-}
-SELECT 1 as test
-}`
-        );
-
-        const result = runMainInVm(coreExecutionRequestFromPath(projectDir));
-
-        expect(result.compile.compiledGraph.graphErrors.compilationErrors.length).deep.equals(2);
-        expect(result.compile.compiledGraph.graphErrors.compilationErrors[0].message).deep.equals(
-          `Conflicting "includeDependentAssertions" properties are not allowed. Dependency A has different values set for this property.`
-        );
-      });
-
-      suite("Action configs", () => {
-        test(`When dependOnDependencyAssertions property is set to true, assertions from A are added as dependencies`, () => {
-          fs.writeFileSync(
-            path.join(projectDir, "definitions/actions.yaml"),
-            `
-actions:
-- view:
-    filename: B.sql
-    dependOnDependencyAssertions: true
-    dependencyTargets:
-      - name: A
-- operation:
-    filename: C.sql
-    dependOnDependencyAssertions: true
-    dependencyTargets:
-      - name: A
-- notebook:
-    filename: notebook.ipynb
-    dependOnDependencyAssertions: true
-    dependencyTargets:
-      - name: A
-`
-          );
-
-          const result = runMainInVm(coreExecutionRequestFromPath(projectDir));
-          expect(
-            asPlainObject(
-              result.compile.compiledGraph.operations.find(
-                operation =>
-                  operation.target.name === prefixAdjustedName(testConfig.namePrefix, "C")
-              ).dependencyTargets.length
-            )
-          ).deep.equals(3);
-          expect(
-            asPlainObject(
-              result.compile.compiledGraph.tables.find(
-                table => table.target.name === prefixAdjustedName(testConfig.namePrefix, "B")
-              ).dependencyTargets.length
-            )
-          ).deep.equals(3);
-          expect(
-            asPlainObject(
-              result.compile.compiledGraph.notebooks.find(
-                notebook =>
-                  notebook.target.name === prefixAdjustedName(testConfig.namePrefix, "notebook")
-              ).dependencyTargets.length
-            )
-          ).deep.equals(3);
-          expect(result.compile.compiledGraph.graphErrors.compilationErrors).deep.equals([]);
-        });
-
-        test(`Setting includeDependentAssertions to true in config.dependencies adds assertions from that dependency to dependencyTargets`, () => {
-          fs.writeFileSync(
-            path.join(projectDir, "definitions/actions.yaml"),
-            `
-actions:
-- view:
-    filename: B.sql
-    dependencyTargets:
-      - name: A
-        includeDependentAssertions: true 
-- operation:
-    filename: C.sql
-    dependencyTargets:
-      - name: A
-        includeDependentAssertions: true
-- notebook:
-    filename: notebook.ipynb
-    dependencyTargets:
-      - name: A
-        includeDependentAssertions: true
-`
-          );
-
-          const result = runMainInVm(coreExecutionRequestFromPath(projectDir));
-
-          expect(result.compile.compiledGraph.graphErrors.compilationErrors).deep.equals([]);
-          expect(
-            asPlainObject(
-              result.compile.compiledGraph.operations.find(
-                operation =>
-                  operation.target.name === prefixAdjustedName(testConfig.namePrefix, "C")
-              ).dependencyTargets.length
-            )
-          ).deep.equals(3);
-          expect(
-            asPlainObject(
-              result.compile.compiledGraph.tables.find(
-                table => table.target.name === prefixAdjustedName(testConfig.namePrefix, "B")
-              ).dependencyTargets.length
-            )
-          ).deep.equals(3);
-          expect(
-            asPlainObject(
-              result.compile.compiledGraph.notebooks.find(
-                notebook =>
-                  notebook.target.name === prefixAdjustedName(testConfig.namePrefix, "notebook")
-              ).dependencyTargets.length
-            )
-          ).deep.equals(3);
-        });
-
-        test(`When dependOnDependencyAssertions=true and includeDependentAssertions=false, the assertions related to dependency should not be added to dependencyTargets`, () => {
-          fs.writeFileSync(
-            path.join(projectDir, "definitions/actions.yaml"),
-            `
-actions:
-- view:
-    filename: B.sql
-    dependOnDependencyAssertions: true
-    dependencyTargets:
-      - name: A
-        includeDependentAssertions: false
-- assertion:
-    filename: B_assert.sql
-    dependencyTargets:
-      - name: B
-- operation:
-    filename: C.sql
-    dependOnDependencyAssertions: true
-    dependencyTargets:
-      - name: A
-        includeDependentAssertions: false
-- notebook:
-    filename: notebook.ipynb
-    dependOnDependencyAssertions: true
-    dependencyTargets:
-      - name: A
-        includeDependentAssertions: false
-      - name: B
-`
-          );
-          fs.writeFileSync(path.join(projectDir, "definitions/B_assert.sql"), "SELECT test from B");
-
-          const result = runMainInVm(coreExecutionRequestFromPath(projectDir));
-
-          expect(result.compile.compiledGraph.graphErrors.compilationErrors).deep.equals([]);
-          expect(
-            asPlainObject(
-              result.compile.compiledGraph.operations.find(
-                operation =>
-                  operation.target.name === prefixAdjustedName(testConfig.namePrefix, "C")
-              ).dependencyTargets.length
-            )
-          ).deep.equals(1);
-          expect(
-            asPlainObject(
-              result.compile.compiledGraph.tables.find(
-                table => table.target.name === prefixAdjustedName(testConfig.namePrefix, "B")
-              ).dependencyTargets.length
-            )
-          ).deep.equals(1);
-          expect(
-            asPlainObject(
-              result.compile.compiledGraph.notebooks.find(
-                notebook =>
-                  notebook.target.name === prefixAdjustedName(testConfig.namePrefix, "notebook")
-              ).dependencyTargets.length
-            )
-          ).deep.equals(3);
-        });
-
-        test(`When dependOnDependencyAssertions=false and includeDependentAssertions=true, the assertions related to dependency should be added to dependencyTargets`, () => {
-          fs.writeFileSync(
-            path.join(projectDir, "definitions/actions.yaml"),
-            `
-actions:
-- view:
-    filename: B.sql
-    dependOnDependencyAssertions: false
-    dependencyTargets:
-      - name: A
-        includeDependentAssertions: true
-- assertion:
-    filename: B_assert.sql
-    dependencyTargets:
-      - name: B
-- operation:
-    filename: C.sql
-    dependOnDependencyAssertions: false
-    dependencyTargets:
-      - name: A
-        includeDependentAssertions: true
-      - name: B
-- notebook:
-    filename: notebook.ipynb
-    dependOnDependencyAssertions: false
-    dependencyTargets:
-      - name: A
-        includeDependentAssertions: true
-      - name: B
-`
-          );
-          fs.writeFileSync(path.join(projectDir, "definitions/B_assert.sql"), "SELECT test from B");
-
-          const result = runMainInVm(coreExecutionRequestFromPath(projectDir));
-
-          expect(result.compile.compiledGraph.graphErrors.compilationErrors).deep.equals([]);
-          expect(
-            asPlainObject(
-              result.compile.compiledGraph.operations.find(
-                operation =>
-                  operation.target.name === prefixAdjustedName(testConfig.namePrefix, "C")
-              ).dependencyTargets.length
-            )
-          ).deep.equals(4);
-          expect(
-            asPlainObject(
-              result.compile.compiledGraph.tables.find(
-                table => table.target.name === prefixAdjustedName(testConfig.namePrefix, "B")
-              ).dependencyTargets.length
-            )
-          ).deep.equals(3);
-          expect(
-            asPlainObject(
-              result.compile.compiledGraph.notebooks.find(
-                notebook =>
-                  notebook.target.name === prefixAdjustedName(testConfig.namePrefix, "notebook")
-              ).dependencyTargets.length
-            )
-          ).deep.equals(4);
-        });
-
-        test(`When includeDependentAssertions property in config and ref are set differently for the same dependency, compilation error is thrown.`, () => {
-          fs.writeFileSync(
-            path.join(projectDir, "definitions/actions.yaml"),
-            `
-actions:
-- view:
-    filename: B.sql
-    dependOnDependencyAssertions: true
-    dependencyTargets:
-      - name: A
-- operation:
-    filename: C.sql
-    dependencyTargets:
-      - name: A
-        includeDependentAssertions: true
-      - name: B
-      - name: A
-        includeDependentAssertions: false
-`
-          );
-          fs.writeFileSync(path.join(projectDir, "definitions/B_assert.sql"), "SELECT test from B");
-
-          const result = runMainInVm(coreExecutionRequestFromPath(projectDir));
-
-          expect(result.compile.compiledGraph.graphErrors.compilationErrors.length).deep.equals(1);
-          expect(result.compile.compiledGraph.graphErrors.compilationErrors[0].message).deep.equals(
-            `Conflicting "includeDependentAssertions" properties are not allowed. Dependency A has different values set for this property.`
-          );
-        });
-      });
-    });
-  });
-
   suite("javascript API", () => {
     suite("publish", () => {
       ["table", "view", "incremental"].forEach(tableType => {
         [
-          TestConfigs.bigqueryWithDefaultProjectAndDataset,
-          { ...TestConfigs.bigqueryWithDatasetSuffix, defaultProject: "defaultProject" },
-          { ...TestConfigs.bigqueryWithNamePrefix, defaultProject: "defaultProject" }
+          WorkflowSettingsTemplates.bigqueryWithDefaultProjectAndDataset,
+          {
+            ...WorkflowSettingsTemplates.bigqueryWithDatasetSuffix,
+            defaultProject: "defaultProject"
+          },
+          { ...WorkflowSettingsTemplates.bigqueryWithNamePrefix, defaultProject: "defaultProject" }
         ].forEach(projectConfig => {
           test(
             `publish for table type ${tableType}, with project suffix ` +
@@ -4487,9 +1174,12 @@ publish("name", {
 
     suite("operate", () => {
       [
-        TestConfigs.bigqueryWithDefaultProjectAndDataset,
-        { ...TestConfigs.bigqueryWithDatasetSuffix, defaultProject: "defaultProject" },
-        { ...TestConfigs.bigqueryWithNamePrefix, defaultProject: "defaultProject" }
+        WorkflowSettingsTemplates.bigqueryWithDefaultProjectAndDataset,
+        {
+          ...WorkflowSettingsTemplates.bigqueryWithDatasetSuffix,
+          defaultProject: "defaultProject"
+        },
+        { ...WorkflowSettingsTemplates.bigqueryWithNamePrefix, defaultProject: "defaultProject" }
       ].forEach(projectConfig => {
         test(
           `operate with project suffix ` +
@@ -4593,9 +1283,12 @@ operate("name", {
 
     suite("assert", () => {
       [
-        TestConfigs.bigqueryWithDefaultProjectAndDataset,
-        { ...TestConfigs.bigqueryWithDatasetSuffix, defaultProject: "defaultProject" },
-        { ...TestConfigs.bigqueryWithNamePrefix, defaultProject: "defaultProject" }
+        WorkflowSettingsTemplates.bigqueryWithDefaultProjectAndDataset,
+        {
+          ...WorkflowSettingsTemplates.bigqueryWithDatasetSuffix,
+          defaultProject: "defaultProject"
+        },
+        { ...WorkflowSettingsTemplates.bigqueryWithNamePrefix, defaultProject: "defaultProject" }
       ].forEach(projectConfig => {
         test(
           `assert with project suffix ` +
@@ -4856,7 +1549,9 @@ publish("name", {
             fs.writeFileSync(
               path.join(projectDir, "workflow_settings.yaml"),
               dumpYaml(
-                dataform.WorkflowSettings.create(TestConfigs.bigqueryWithDefaultProjectAndDataset)
+                dataform.WorkflowSettings.create(
+                  WorkflowSettingsTemplates.bigqueryWithDefaultProjectAndDataset
+                )
               )
             );
             fs.mkdirSync(path.join(projectDir, "definitions"));
@@ -4882,76 +1577,3 @@ publish("name", {type: "${fromType}", schema: "schemaOverride"}).type("${toType}
     });
   });
 });
-
-function coreExecutionRequestFromPath(
-  projectDir: string,
-  projectConfigOverride?: dataform.ProjectConfig
-): dataform.CoreExecutionRequest {
-  return dataform.CoreExecutionRequest.create({
-    compile: {
-      compileConfig: {
-        projectDir,
-        filePaths: walkDirectoryForFilenames(projectDir),
-        projectConfigOverride
-      }
-    }
-  });
-}
-
-// A VM is needed when running main because Node functions like `require` are overridden.
-function runMainInVm(
-  coreExecutionRequest: dataform.CoreExecutionRequest
-): dataform.CoreExecutionResponse {
-  const projectDir = coreExecutionRequest.compile.compileConfig.projectDir;
-
-  // Copy over the build Dataform Core that is set up as a node_modules directory.
-  fs.copySync(`${process.cwd()}/core/node_modules`, `${projectDir}/node_modules`);
-
-  const compiler = compile as CompilerFunction;
-  // Then use vm2's native compiler integration to apply the compiler to files.
-  const nodeVm = new NodeVM({
-    // Inheriting the console makes console.logs show when tests are running, which is useful for
-    // debugging.
-    console: "inherit",
-    wrapper: "none",
-    require: {
-      builtin: ["path"],
-      context: "sandbox",
-      external: true,
-      root: projectDir,
-      resolve: (moduleName, parentDirName) =>
-        path.join(parentDirName, path.relative(parentDirName, projectDir), moduleName)
-    },
-    sourceExtensions: SOURCE_EXTENSIONS,
-    compiler
-  });
-
-  const encodedCoreExecutionRequest = encode64(dataform.CoreExecutionRequest, coreExecutionRequest);
-  const vmIndexFileName = path.resolve(path.join(projectDir, "index.js"));
-  const encodedCoreExecutionResponse = nodeVm.run(
-    `return require("@dataform/core").main("${encodedCoreExecutionRequest}")`,
-    vmIndexFileName
-  );
-  return decode64(dataform.CoreExecutionResponse, encodedCoreExecutionResponse);
-}
-
-function walkDirectoryForFilenames(projectDir: string, relativePath: string = ""): string[] {
-  let paths: string[] = [];
-  fs.readdirSync(path.join(projectDir, relativePath), { withFileTypes: true })
-    .filter(directoryEntry => directoryEntry.name !== "node_modules")
-    .forEach(directoryEntry => {
-      if (directoryEntry.isDirectory()) {
-        paths = paths.concat(walkDirectoryForFilenames(projectDir, directoryEntry.name));
-        return;
-      }
-      const fileExtension = directoryEntry.name.split(".").slice(-1)[0];
-      if (directoryEntry.isFile() && SOURCE_EXTENSIONS.includes(fileExtension)) {
-        paths.push(directoryEntry.name);
-      }
-    });
-  return paths.map(filename => path.join(relativePath, filename));
-}
-
-function prefixAdjustedName(prefix: string | undefined, name: string) {
-  return prefix ? `${prefix}_${name}` : name;
-}
